+++ test/unit/org/authInfoTest.ts
@@ -4,367 +4,155 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
+
 /* eslint-disable camelcase */
 /* eslint-disable @typescript-eslint/ban-types */
 
-import * as dns from 'dns';
 import * as pathImport from 'path';
-import * as fs from 'fs';
-import { cloneJson, env, includes, set } from '@salesforce/kit';
-import { spyMethod, stubMethod } from '@salesforce/ts-sinon';
+import * as dns from 'dns';
+import * as jwt from 'jsonwebtoken';
+import { cloneJson, env, includes } from '@salesforce/kit';
+import { stubMethod, spyMethod } from '@salesforce/ts-sinon';
 import { AnyJson, ensureString, getJsonMap, getString, JsonMap, toJsonMap } from '@salesforce/ts-types';
 import { assert, expect } from 'chai';
-import { OAuth2, OAuth2Config } from 'jsforce';
-import { match } from 'sinon';
 import { Transport } from 'jsforce/lib/transport';
-import * as jwt from 'jsonwebtoken';
-import { AuthFields, AuthInfo } from '../../../src/org';
-import { Config } from '../../../src/config/config';
-import { ConfigAggregator } from '../../../src/config/configAggregator';
-import { ConfigFile } from '../../../src/config/configFile';
-import { ConfigContents } from '../../../src/config/configStore';
-import { AliasAccessor, OrgAccessor } from '../../../src/stateAggregator';
-import { Crypto } from '../../../src/crypto/crypto';
-import { SfError } from '../../../src/sfError';
+
+import { OAuth2 } from 'jsforce';
+import { SinonSpy, SinonStub } from 'sinon';
+import { AuthFields, AuthInfo, OAuth2Config } from '../../../src/org';
 import { MockTestOrgData, testSetup } from '../../../src/testSetup';
 import { OrgConfigProperties } from '../../../src/org/orgConfigProperties';
-import { AuthInfoConfig } from '../../../src/config/authInfoConfig';
+import { AliasAccessor, OrgAccessor } from '../../../src/stateAggregator';
+import { Crypto } from '../../../src/crypto/crypto';
+import { Config } from '../../../src/config/config';
+import { SfdcUrl } from '../../../src/util/sfdcUrl';
 
-const TEST_KEY = {
-  service: 'sfdx',
-  account: 'local',
-  key: '8e8fd1e6dc06a37bf420898dbc3ee35c',
-};
-
-// Setup the test environment.
-const $$ = testSetup();
-
-describe('AuthInfo No fs mock', () => {
-  beforeEach(() => {
-    // Testing crypto functionality, so restore global stubs.
-    $$.SANDBOXES.CRYPTO.restore();
-    $$.SANDBOXES.CONFIG.restore();
-    $$.SANDBOXES.ORGS.restore();
-
-    stubMethod($$.SANDBOX, Crypto.prototype, 'getKeyChain').callsFake(() =>
-      Promise.resolve({
-        setPassword: () => Promise.resolve(),
-        getPassword: (data: JsonMap, cb: (val1: AnyJson, key: string) => {}) => cb(null, TEST_KEY.key),
-      })
-    );
-  });
-
-  it('missing config', async () => {
-    const expectedErrorName = 'NamedOrgNotFoundError';
-    try {
-      await AuthInfo.create({ username: 'does_not_exist@gb.com' });
-      assert.fail(`should have thrown error with name: ${expectedErrorName}`);
-    } catch (e) {
-      expect(e).to.have.property('name', expectedErrorName);
-    }
-  });
-
-  it('invalid devhub username', async () => {
-    const expectedErrorName = 'NamedOrgNotFoundError';
-    try {
-      await AuthInfo.create({ username: 'does_not_exist@gb.com', isDevHub: true });
-      assert.fail(`should have thrown error with name: ${expectedErrorName}`);
-    } catch (e) {
-      expect(e).to.have.property('name', expectedErrorName);
-      expect(e).to.have.property('message', 'No authorization information found for does_not_exist@gb.com.');
-    }
-  });
-});
-
-// Cleanly encapsulate the test data.
-class MetaAuthDataMock {
-  private _instanceUrl = 'https://mydevhub.localhost.internal.salesforce.com:6109';
-  private _accessToken = 'authInfoTest_access_token';
-  private _encryptedAccessToken: string = this._accessToken;
-  private _refreshToken = 'authInfoTest_refresh_token';
-  private _encryptedRefreshToken: string = this._refreshToken;
-  private _clientId = 'authInfoTest_client_id';
-  private _loginUrl = 'https://foo.bar.baz';
-  private _jwtUsername = 'authInfoTest_username_JWT';
-  private _redirectUri = 'http://localhost:1717/OauthRedirect';
-  private _authCode = 'authInfoTest_authCode';
-  private _authInfoLookupCount = 0;
-  private _defaultConnectedAppInfo: AuthFields = {
+class AuthInfoMockOrg extends MockTestOrgData {
+  public privateKey = 'authInfoTest/jwt/server.key';
+  public expirationDate = '12-02-20';
+  public encryptedAccessToken = this.accessToken;
+  public defaultConnectedAppInfo = {
     clientId: 'SalesforceDevelopmentExperience',
     clientSecret: '1384510088588713504',
   };
-  private _expirationDate = '12-02-20';
-  private _clientSecret = 'client_secret';
-  private _orgId = 'testOrgId';
 
-  public constructor() {
-    this._jwtUsername = `${this._jwtUsername}_${$$.uniqid()}`;
-  }
-
-  public get instanceUrl(): string {
-    return this._instanceUrl;
-  }
-
-  public set instanceUrl(value: string) {
-    this._instanceUrl = value;
-  }
-
-  public get accessToken(): string {
-    return this._accessToken;
-  }
-
-  public get refreshToken(): string {
-    return this._refreshToken;
-  }
-
-  public get clientId(): string {
-    return this._clientId;
-  }
-
-  public get loginUrl(): string {
-    return this._loginUrl;
-  }
-
-  public set loginUrl(value: string) {
-    this._loginUrl = value;
-  }
-
-  public get jwtUsername(): string {
-    return this._jwtUsername;
-  }
-
-  public set jwtUsername(value: string) {
-    this._jwtUsername = value;
-  }
-
-  public get username(): string {
-    return this._jwtUsername;
-  }
-
-  public get redirectUri(): string {
-    return this._redirectUri;
-  }
-
-  public get authCode(): string {
-    return this._authCode;
-  }
-
-  public set authCode(value: string) {
-    this._authCode = value;
-  }
-
-  public get defaultConnectedAppInfo(): AuthFields {
-    return this._defaultConnectedAppInfo;
-  }
-
-  public get encryptedAccessToken(): string {
-    return this._encryptedAccessToken;
-  }
-
-  public set encryptedAccessToken(value: string) {
-    this._encryptedAccessToken = value;
-  }
-
-  public get encryptedRefreshToken(): string {
-    return this._encryptedRefreshToken;
-  }
-
-  public set encryptedRefreshToken(value: string) {
-    this._encryptedRefreshToken = value;
-  }
-
-  public get expirationDate(): string {
-    return this._expirationDate;
-  }
-
-  public set expirationDate(value: string) {
-    this._expirationDate = value;
-  }
-
-  public get authInfoLookupCount(): number {
-    return this._authInfoLookupCount;
-  }
-
-  public get clientSecret(): string {
-    return this._clientSecret;
-  }
-
-  public get orgId(): string {
-    return this._orgId;
-  }
-
-  public fetchConfigInfo(path: string): ConfigContents {
-    if (path.toUpperCase().includes('JWT')) {
-      this._authInfoLookupCount = this._authInfoLookupCount + 1;
-      const configContents = {};
-
-      set(configContents, 'instanceUrl', 'https://mydevhub.localhost.internal.salesforce.com:6109');
-      set(configContents, 'accessToken', this.encryptedAccessToken);
-      set(configContents, 'privateKey', '123456');
-      set(configContents, 'username', this.username);
-      return configContents;
-    } else {
-      return {};
-    }
-  }
-
-  public async statForKeyFile(path: string): Promise<{}> {
-    if (!path.includes('key.json')) {
-      return new SfError(`Unexpected path: ${path}`, 'UnexpectedInput');
-    }
-
-    return Promise.resolve({
-      dev: 16777221,
-      mode: 16768,
-      nlink: 32,
-      uid: 1613127851,
-      gid: 0,
-      rdev: 0,
-      blksize: 4194304,
-      ino: 81943357,
-      size: 1024,
-      blocks: 0,
-      atimeMs: 1517934734270.9426,
-      mtimeMs: 1517879310026.148,
-      ctimeMs: 1517879310026.148,
-      birthtimeMs: 1510678165000,
-      atime: new Date('2018-02-06T16:32:14.271Z'),
-      mtime: new Date('2018-02-06T01:08:30.026Z'),
-      ctime: new Date('2018-02-06T01:08:30.026Z'),
-      birthtime: new Date('2017-11-14T16:49:25.000Z'),
-    });
+  public async getConfig(): Promise<AuthFields> {
+    return {
+      accessToken: this.accessToken,
+      clientId: this.clientId,
+      instanceUrl: this.instanceUrl,
+      loginUrl: this.loginUrl,
+      privateKey: this.privateKey,
+      username: this.username,
+      orgId: this.orgId,
+    };
   }
 }
 
 describe('AuthInfo', () => {
-  let authInfoInit: sinon.SinonSpy;
-  let authInfoUpdate: sinon.SinonSpy;
-  let authInfoBuildJwtConfig: sinon.SinonSpy;
-  let authInfoBuildRefreshTokenConfig: sinon.SinonSpy;
-  let authInfoExchangeToken: sinon.SinonSpy;
+  // Setup the test environment.
+  const $$ = testSetup();
 
-  let configFileWrite: sinon.SinonStub;
+  let testOrg: AuthInfoMockOrg;
 
-  let readFileStub: sinon.SinonStub;
-  let _postParmsStub: sinon.SinonStub;
+  let postParamsStub: SinonStub;
+  let orgAccessorReadSpy: SinonSpy;
+  let authInfoStubs = {} as Record<string, SinonSpy | SinonStub>;
 
-  let testMetadata: MetaAuthDataMock;
+  // Walk an object deeply looking for the attribute name of clientSecret or values that contain the client secret
+  // or decrypted refresh token.
+  const walkAndSearchForSecrets = (obj: JsonMap, decryptedRefreshToken: string) => {
+    const keys = Object.keys(obj);
+    keys.forEach((key: string) => {
+      const child = getJsonMap(obj, key);
+      if (child) {
+        walkAndSearchForSecrets(child, decryptedRefreshToken);
+      }
+      const keyUpper = key.toUpperCase();
+
+      // If the key is likely a clientSecret "ish" attribute and the value is a string.
+      // reminder:'clientSecretFn' is always legit.
+      if (keyUpper.includes('SECRET') && keyUpper.includes('CLIENT') && getString(obj, key)) {
+        throw new Error('Key indicates client secret.');
+      }
+
+      if (includes(getJsonMap(obj, key), testOrg.defaultConnectedAppInfo.clientSecret)) {
+        throw new Error(`Client secret present as value in object with key: ${key}`);
+      }
+
+      if (includes(getJsonMap(obj, key), decryptedRefreshToken)) {
+        throw new Error(`Refresh token present as value in object with key: ${key}`);
+      }
+    });
+  };
+
+  const stubUserRequest = (
+    userInfoResponse: JsonMap = {
+      statusCode: 200,
+      body: { preferred_username: testOrg.username, organization_id: testOrg.orgId },
+    },
+    userResponse: JsonMap = { statusCode: 200, body: { Username: testOrg.username.toUpperCase() } }
+  ): SinonStub => {
+    const userInfoResponseBody = {
+      statusCode: userInfoResponse.statusCode,
+      body: JSON.stringify(userInfoResponse.body),
+    };
+    const userResponseBody = {
+      statusCode: userResponse.statusCode,
+      body: JSON.stringify(userResponse.body),
+    };
+    return stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest')
+      .onFirstCall()
+      .resolves(userInfoResponseBody)
+      .onSecondCall()
+      .resolves(userResponseBody);
+  };
 
   beforeEach(async () => {
-    // Testing config functionality, so restore global stubs.
-    $$.SANDBOXES.CONFIG.restore();
+    testOrg = new AuthInfoMockOrg();
 
-    testMetadata = new MetaAuthDataMock();
+    postParamsStub = stubMethod($$.SANDBOX, OAuth2.prototype, '_postParams');
 
-    stubMethod($$.SANDBOX, fs, 'stat').callsFake(async (path: string) => {
-      return testMetadata.statForKeyFile(path);
-    });
+    orgAccessorReadSpy = spyMethod($$.SANDBOX, OrgAccessor.prototype, 'read');
 
-    // Common stubs
-    configFileWrite = stubMethod($$.SANDBOX, ConfigFile.prototype, 'write').callsFake(async () => {
-      return Promise.resolve();
-    });
-
-    stubMethod($$.SANDBOX, fs.promises, 'writeFile')
-      .withArgs(match(/.*key.json/))
-      .resolves()
-      .rejects(); // .callThrough;
-    readFileStub = stubMethod($$.SANDBOX, fs.promises, 'readFile')
-      .withArgs(match(/.*key.json/))
-      .resolves({})
-      .rejects();
-
-    function read(this: AuthInfoConfig) {
-      this.setContentsFromObject(testMetadata.fetchConfigInfo(this.getPath()));
-      return this.getContents();
-    }
-
-    stubMethod($$.SANDBOX, ConfigFile.prototype, 'read').callsFake(read);
-    stubMethod($$.SANDBOX, ConfigFile.prototype, 'readSync').callsFake(read);
-
-    const crypto = await Crypto.create();
-    testMetadata.encryptedAccessToken = crypto.encrypt(testMetadata.accessToken) || '';
-    testMetadata.encryptedRefreshToken = crypto.encrypt(testMetadata.refreshToken) || '';
-
-    // These stubs return different objects based on the tests
-    _postParmsStub = stubMethod($$.SANDBOX, OAuth2.prototype, '_postParams');
-
-    // Spies
-    authInfoInit = spyMethod($$.SANDBOX, AuthInfo.prototype, 'initAuthOptions');
-    authInfoUpdate = spyMethod($$.SANDBOX, AuthInfo.prototype, 'update');
-    authInfoBuildJwtConfig = spyMethod($$.SANDBOX, AuthInfo.prototype, 'authJwt');
-    authInfoBuildRefreshTokenConfig = spyMethod($$.SANDBOX, AuthInfo.prototype, 'buildRefreshTokenConfig');
-    authInfoExchangeToken = spyMethod($$.SANDBOX, AuthInfo.prototype, 'exchangeToken');
+    authInfoStubs = {
+      initAuthOptions: spyMethod($$.SANDBOX, AuthInfo.prototype, 'initAuthOptions'),
+      update: spyMethod($$.SANDBOX, AuthInfo.prototype, 'update'),
+      authJwt: spyMethod($$.SANDBOX, AuthInfo.prototype, 'authJwt'),
+      buildRefreshTokenConfig: spyMethod($$.SANDBOX, AuthInfo.prototype, 'buildRefreshTokenConfig'),
+      exchangeToken: spyMethod($$.SANDBOX, AuthInfo.prototype, 'exchangeToken'),
+    };
   });
 
   describe('Secret Tests', () => {
     let authInfo: AuthInfo;
     let decryptedRefreshToken: string;
+
     beforeEach(async () => {
-      const authCodeConfig = {
-        authCode: testMetadata.authCode,
-        loginUrl: testMetadata.loginUrl,
-      };
+      const oauth2Options = { authCode: testOrg.authcode, loginUrl: testOrg.loginUrl };
       const authResponse = {
-        access_token: testMetadata.accessToken,
-        instance_url: testMetadata.instanceUrl,
+        access_token: testOrg.accessToken,
+        instance_url: testOrg.instanceUrl,
         id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
-        refresh_token: testMetadata.refreshToken,
+        refresh_token: testOrg.refreshToken,
       };
 
       // Stub the http requests (OAuth2.requestToken() and the request for the username)
-      _postParmsStub.returns(Promise.resolve(authResponse));
-      const userInfoResponseBody = {
-        body: JSON.stringify({ preferred_username: testMetadata.username, organization_id: testMetadata.orgId }),
-      };
-      const userResponseBody = {
-        body: JSON.stringify({ Username: testMetadata.username.toUpperCase() }),
-      };
-      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest')
-        .onFirstCall()
-        .returns(Promise.resolve(userInfoResponseBody))
-        .onSecondCall()
-        .returns(Promise.resolve(userResponseBody));
-      authInfo = await AuthInfo.create({ oauth2Options: authCodeConfig });
+      postParamsStub.resolves(authResponse);
+      stubUserRequest();
+      authInfo = await AuthInfo.create({ oauth2Options });
 
       decryptedRefreshToken = authInfo.getFields(true).refreshToken;
     });
-
-    // Walk an object deeply looking for the attribute name of clientSecret or values that contain the client secret
-    // or decrypted refresh token.
-    const walkAndSearchForSecrets = (obj: JsonMap) => {
-      const keys = Object.keys(obj);
-      keys.forEach((key: string) => {
-        const child = getJsonMap(obj, key);
-        if (child) {
-          walkAndSearchForSecrets(child);
-        }
-        const keyUpper = key.toUpperCase();
-
-        // If the key is likely a clientSecret "ish" attribute and the value is a string.
-        // reminder:'clientSecretFn' is always legit.
-        if (keyUpper.includes('SECRET') && keyUpper.includes('CLIENT') && getString(obj, key)) {
-          throw new Error('Key indicates client secret.');
-        }
-
-        if (includes(getJsonMap(obj, key), testMetadata.defaultConnectedAppInfo.clientSecret)) {
-          throw new Error(`Client secret present as value in object with key: ${key}`);
-        }
-
-        if (includes(getJsonMap(obj, key), decryptedRefreshToken)) {
-          throw new Error(`Refresh token present as value in object with key: ${key}`);
-        }
-      });
-    };
 
     describe('getFields', () => {
       it('return value should not have a client secret or decrypted refresh token', () => {
         const fields = authInfo.getFields();
         const strObj = JSON.stringify(fields);
         // verify the returned object doesn't have secrets
-        expect(() => walkAndSearchForSecrets(toJsonMap(fields) || {})).to.not.throw();
+        expect(() => walkAndSearchForSecrets(toJsonMap(fields) || {}, decryptedRefreshToken)).to.not.throw();
 
-        expect(strObj).does.not.include(ensureString(testMetadata.defaultConnectedAppInfo.clientSecret));
+        expect(strObj).does.not.include(ensureString(testOrg.defaultConnectedAppInfo.clientSecret));
         expect(strObj).does.not.include(decryptedRefreshToken);
       });
     });
@@ -381,14 +169,14 @@ describe('AuthInfo', () => {
 
     describe('getConnectionOptions', () => {
       it('return value should not have a client secret or decrypted refresh token', () => {
-        const fields: AuthFields = authInfo.getConnectionOptions();
-        const strObj: string = JSON.stringify(fields);
+        const fields = authInfo.getConnectionOptions();
+        const strObj = JSON.stringify(fields);
 
         // verify the returned object doesn't have secrets
-        expect(() => walkAndSearchForSecrets(toJsonMap(fields) || {})).to.not.throw();
+        expect(() => walkAndSearchForSecrets(toJsonMap(fields) || {}, decryptedRefreshToken)).to.not.throw();
 
         // double check the stringified objects don't have secrets.
-        expect(strObj).does.not.include(ensureString(testMetadata.defaultConnectedAppInfo.clientSecret));
+        expect(strObj).does.not.include(ensureString(testOrg.defaultConnectedAppInfo.clientSecret));
         expect(strObj).does.not.include(decryptedRefreshToken);
       });
     });
@@ -398,19 +186,43 @@ describe('AuthInfo', () => {
         const authInfoString = JSON.stringify(authInfo);
 
         // verify the returned object doesn't have secrets
-        expect(() => walkAndSearchForSecrets(toJsonMap(authInfo) || {})).to.not.throw();
+        expect(() => walkAndSearchForSecrets(toJsonMap(authInfo) || {}, decryptedRefreshToken)).to.not.throw();
 
         // double check the stringified objects don't have secrets.
-        expect(authInfoString).does.not.include(ensureString(testMetadata.defaultConnectedAppInfo.clientSecret));
+        expect(authInfoString).does.not.include(ensureString(testOrg.defaultConnectedAppInfo.clientSecret));
         expect(authInfoString).does.not.include(decryptedRefreshToken);
       });
     });
   });
 
-  describe('create()', () => {
+  describe('create', () => {
+    const verifyAuthInfoRefreshToken = (
+      authInfo: AuthInfo,
+      authResponse: { access_token: string; instance_url: string },
+      config?: { clientId?: string }
+    ) => {
+      // Verify the returned AuthInfo instance
+      const authInfoConnOpts = authInfo.getConnectionOptions();
+      expect(authInfoConnOpts).to.have.property('accessToken', authResponse.access_token);
+      expect(authInfoConnOpts).to.have.property('instanceUrl', authResponse.instance_url);
+      expect(authInfoConnOpts).to.not.have.property('refreshToken');
+      expect(authInfoConnOpts['oauth2']).to.have.property('loginUrl', testOrg.instanceUrl);
+      if (config?.clientId) {
+        expect(authInfoConnOpts['oauth2']).to.have.property('clientId', config?.clientId);
+      } else {
+        expect(authInfoConnOpts['oauth2']).to.have.property('clientId', testOrg.defaultConnectedAppInfo.clientId);
+      }
+
+      expect(authInfoConnOpts['oauth2']).to.have.property('redirectUri', testOrg.redirectUri);
+      expect(authInfo.getUsername()).to.equal(testOrg.username);
+      expect(authInfo.isAccessTokenFlow(), 'authInfo.isAccessTokenFlow() should be false').to.be.false;
+      expect(authInfo.isRefreshTokenFlow(), 'authInfo.isRefreshTokenFlow() should be true').to.be.true;
+      expect(authInfo.isJwt(), 'authInfo.isJwt() should be false').to.be.false;
+      expect(authInfo.isOauth(), 'authInfo.isOauth() should be true').to.be.true;
+    };
+
     it('should return an AuthInfo instance when passed an access token as username', async () => {
-      stubMethod($$.SANDBOX, ConfigAggregator.prototype, 'loadProperties').callsFake(async () => {});
-      stubMethod($$.SANDBOX, ConfigAggregator.prototype, 'getPropertyValue').returns(testMetadata.instanceUrl);
+      $$.stubConfig({ [OrgConfigProperties.ORG_INSTANCE_URL]: testOrg.instanceUrl });
 
       const username =
         '00Dxx0000000001!AQEAQI3AIbublfW11ATFJl9T122vVPj5QaInBp6h9nPsUK8oW4rW5Os0ZjtsUU.DG9rXytUCh3RZvc_XYoRULiHeTMjyi6T1';
@@ -418,8 +230,8 @@ describe('AuthInfo', () => {
 
       const expectedFields = {
         accessToken: username,
-        instanceUrl: testMetadata.instanceUrl,
-        loginUrl: testMetadata.instanceUrl,
+        instanceUrl: testOrg.instanceUrl,
+        loginUrl: testOrg.instanceUrl,
       };
       expect(authInfo.getConnectionOptions()).to.deep.equal(expectedFields);
       expect(authInfo.isAccessTokenFlow(), 'authInfo.isAccessTokenFlow() should be true').to.be.true;
@@ -429,17 +241,16 @@ describe('AuthInfo', () => {
     });
 
     it('should return an AuthInfo instance when passed a parent username', async () => {
-      stubMethod($$.SANDBOX, ConfigAggregator.prototype, 'loadProperties').callsFake(async () => {});
-      stubMethod($$.SANDBOX, ConfigAggregator.prototype, 'getPropertyValue').returns(testMetadata.instanceUrl);
+      $$.stubConfig({ [OrgConfigProperties.ORG_INSTANCE_URL]: testOrg.instanceUrl });
       // Stub the http request (OAuth2.refreshToken())
-      // This will be called for both, and we want to make sure the clientSecrete is the
+      // This will be called for both, and we want to make sure the clientSecret is the
       // same for both.
-      _postParmsStub.callsFake((params) => {
-        expect(params.client_secret).to.deep.equal(testMetadata.clientSecret);
+      postParamsStub.callsFake((params) => {
+        expect(params.client_secret).to.deep.equal(testOrg.clientSecret);
         return {
-          access_token: testMetadata.accessToken,
-          instance_url: testMetadata.instanceUrl,
-          refresh_token: testMetadata.refreshToken,
+          access_token: testOrg.accessToken,
+          instance_url: testOrg.instanceUrl,
+          refresh_token: testOrg.refreshToken,
           id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
         };
       });
@@ -448,56 +259,46 @@ describe('AuthInfo', () => {
       await AuthInfo.create({
         username: parentUsername,
         oauth2Options: {
-          clientId: testMetadata.clientId,
-          clientSecret: testMetadata.clientSecret,
-          loginUrl: testMetadata.instanceUrl,
-          authCode: testMetadata.authCode,
+          clientId: testOrg.clientId,
+          clientSecret: testOrg.clientSecret,
+          loginUrl: testOrg.instanceUrl,
+          authCode: testOrg.authcode,
         },
       });
 
       const authInfo = await AuthInfo.create({
-        username: testMetadata.username,
+        username: testOrg.username,
         parentUsername,
         oauth2Options: {
-          loginUrl: testMetadata.instanceUrl,
-          authCode: testMetadata.authCode,
+          loginUrl: testOrg.instanceUrl,
+          authCode: testOrg.authcode,
         },
       });
 
-      expect(_postParmsStub.calledTwice).to.true;
+      expect(postParamsStub.calledTwice).to.true;
       expect(authInfo.isAccessTokenFlow(), 'authInfo.isAccessTokenFlow() should be false').to.be.false;
       expect(authInfo.isRefreshTokenFlow(), 'authInfo.isRefreshTokenFlow() should be false').to.be.true;
       expect(authInfo.isJwt(), 'authInfo.isJwt() should be false').to.be.false;
       expect(authInfo.isOauth(), 'authInfo.isOauth() should be true').to.be.true;
 
       const expectedAuthConfig = {
-        accessToken: testMetadata.accessToken,
-        instanceUrl: testMetadata.instanceUrl,
-        username: testMetadata.username,
+        accessToken: testOrg.accessToken,
+        instanceUrl: testOrg.instanceUrl,
+        username: testOrg.username,
         orgId: '00DAuthInfoTest_orgId',
-        loginUrl: testMetadata.instanceUrl,
-        refreshToken: testMetadata.refreshToken,
-        clientId: testMetadata.clientId,
-        clientSecret: testMetadata.clientSecret,
+        loginUrl: testOrg.instanceUrl,
+        refreshToken: testOrg.refreshToken,
+        clientId: testOrg.clientId,
+        clientSecret: testOrg.clientSecret,
         isDevHub: false,
       };
-      expect(authInfoUpdate.secondCall.args[0]).to.deep.equal(expectedAuthConfig);
+      expect(authInfoStubs.update.secondCall.args[0]).to.deep.equal(expectedAuthConfig);
     });
 
     it('should return an AuthInfo instance when passed an access token and instanceUrl for the access token flow', async () => {
-      const userInfoResponseBody = {
-        body: JSON.stringify({ preferred_username: testMetadata.username, organization_id: testMetadata.orgId }),
-      };
-      const userResponseBody = {
-        body: JSON.stringify({ Username: testMetadata.username.toUpperCase() }),
-      };
-      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest')
-        .onFirstCall()
-        .returns(Promise.resolve(userInfoResponseBody))
-        .onSecondCall()
-        .returns(Promise.resolve(userResponseBody));
-      stubMethod($$.SANDBOX, ConfigAggregator.prototype, 'loadProperties').callsFake(async () => {});
-      stubMethod($$.SANDBOX, ConfigAggregator.prototype, 'getPropertyValue').returns(testMetadata.instanceUrl);
+      $$.stubConfig({ [OrgConfigProperties.ORG_INSTANCE_URL]: testOrg.instanceUrl });
+
+      stubUserRequest();
 
       const accessToken =
         '00Dxx0000000001!AQEAQI3AIbublfW11ATFJl9T122vVPj5QaInBp6h9nPsUK8oW4rW5Os0ZjtsUU.DG9rXytUCh3RZvc_XYoRULiHeTMjyi6T1';
@@ -505,15 +306,15 @@ describe('AuthInfo', () => {
         username: 'test',
         accessTokenOptions: {
           accessToken,
-          instanceUrl: testMetadata.instanceUrl,
-          loginUrl: testMetadata.instanceUrl,
+          instanceUrl: testOrg.instanceUrl,
+          loginUrl: testOrg.instanceUrl,
         },
       });
 
       const expectedFields = {
         accessToken,
-        instanceUrl: testMetadata.instanceUrl,
-        loginUrl: testMetadata.instanceUrl,
+        instanceUrl: testOrg.instanceUrl,
+        loginUrl: testOrg.instanceUrl,
       };
       expect(authInfo.getConnectionOptions()).to.deep.equal(expectedFields);
       expect(authInfo.isAccessTokenFlow(), 'authInfo.isAccessTokenFlow() should be true').to.be.true;
@@ -522,34 +323,26 @@ describe('AuthInfo', () => {
       expect(authInfo.isOauth(), 'authInfo.isOauth() should be false').to.be.false;
     });
 
-    //
-    // JWT Tests
-    //
-
-    describe('ordered test', () => {
-      // There is an implicit order in these tests. Hence the isolation in the "describe" and the unique
-      // username that is generated in the MetaMock constructor.
-      const sharedTestMeta = new MetaAuthDataMock();
-      beforeEach(async () => {
-        testMetadata = sharedTestMeta;
-      });
-
+    describe('JWT', () => {
       it('should return a JWT AuthInfo instance when passed a username and JWT auth options', async () => {
+        $$.setConfigStubContents('AuthInfoConfig', { contents: await testOrg.getConfig() });
         const jwtConfig = {
-          clientId: testMetadata.clientId,
-          loginUrl: testMetadata.loginUrl,
-          privateKey: 'authInfoTest/jwt/server.key',
+          clientId: testOrg.clientId,
+          loginUrl: testOrg.loginUrl,
+          privateKey: testOrg.privateKey,
         };
         const jwtConfigClone = cloneJson(jwtConfig);
         const authResponse = {
-          access_token: testMetadata.accessToken,
-          instance_url: testMetadata.instanceUrl,
+          access_token: testOrg.accessToken,
+          instance_url: testOrg.instanceUrl,
           id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
         };
 
         // Stub file I/O, http requests, and the DNS lookup
-        readFileStub.resolves('authInfoTest_private_key');
-        _postParmsStub.resolves(authResponse);
+        const readJwtKey = stubMethod($$.SANDBOX, AuthInfo.prototype, 'readJwtKey').resolves(
+          'authInfoTest_private_key'
+        );
+        postParamsStub.resolves(authResponse);
         stubMethod($$.SANDBOX, jwt, 'sign').resolves('authInfoTest_jwtToken');
         stubMethod($$.SANDBOX, dns, 'lookup').callsFake((url: string, done: (v: AnyJson, w: JsonMap) => {}) =>
           done(null, { address: '1.1.1.1', family: 4 })
@@ -557,7 +350,7 @@ describe('AuthInfo', () => {
 
         // Create the JWT AuthInfo instance
         const authInfo = await AuthInfo.create({
-          username: testMetadata.jwtUsername,
+          username: testOrg.username,
           oauth2Options: jwtConfig,
         });
 
@@ -566,677 +359,672 @@ describe('AuthInfo', () => {
         expect(authInfoConnOpts).to.have.property('accessToken', authResponse.access_token);
         expect(authInfoConnOpts).to.have.property('instanceUrl', authResponse.instance_url);
         expect(authInfoConnOpts).to.have.property('refreshFn').and.is.a('function');
-        expect(authInfo.getUsername()).to.equal(testMetadata.jwtUsername);
+        expect(authInfo.getUsername()).to.equal(testOrg.username);
         expect(authInfo.isAccessTokenFlow(), 'authInfo.isAccessTokenFlow() should be false').to.be.false;
         expect(authInfo.isRefreshTokenFlow(), 'authInfo.isRefreshTokenFlow() should be false').to.be.false;
         expect(authInfo.isJwt(), 'authInfo.isJwt() should be true').to.be.true;
         expect(authInfo.isOauth(), 'authInfo.isOauth() should be false').to.be.false;
 
         // Verify expected methods are called with expected args
-        expect(authInfoInit.called).to.be.true;
-        expect(authInfoInit.firstCall.args[0]).to.equal(jwtConfig);
-        expect(authInfoUpdate.called).to.be.true;
-        expect(authInfoBuildJwtConfig.called).to.be.true;
-        expect(authInfoBuildJwtConfig.firstCall.args[0]).to.include(jwtConfig);
+        expect(authInfoStubs.initAuthOptions.called).to.be.true;
+        expect(authInfoStubs.initAuthOptions.firstCall.args[0]).to.equal(jwtConfig);
+        expect(authInfoStubs.update.called).to.be.true;
+        expect(authInfoStubs.authJwt.called).to.be.true;
+        expect(authInfoStubs.authJwt.firstCall.args[0]).to.include(jwtConfig);
         expect(
-          testMetadata.authInfoLookupCount,
+          orgAccessorReadSpy.callCount,
           'should have read an auth file once to ensure auth data did not already exist'
         ).to.equal(1);
-        // expect(readFileStub.called).to.be.true;
+        expect(readJwtKey.called).to.be.true;
 
         // Verify the jwtConfig object was not mutated by init() or authJwt()
         expect(jwtConfig).to.deep.equal(jwtConfigClone);
 
         const expectedAuthConfig = {
           accessToken: authResponse.access_token,
-          clientId: testMetadata.clientId,
-          instanceUrl: testMetadata.instanceUrl,
+          clientId: testOrg.clientId,
+          instanceUrl: testOrg.instanceUrl,
           orgId: authResponse.id.split('/')[0],
           loginUrl: jwtConfig.loginUrl,
           privateKey: jwtConfig.privateKey,
           isDevHub: false,
         };
-        expect(authInfoUpdate.firstCall.args[0]).to.deep.equal(expectedAuthConfig);
+        expect(authInfoStubs.update.firstCall.args[0]).to.deep.equal(expectedAuthConfig);
+      });
+
+      it('should return a JWT AuthInfo instance when passed a username from an auth file', async () => {
+        const jwtData = await testOrg.getConfig();
+        $$.setConfigStubContents('AuthInfoConfig', { contents: jwtData });
+
+        // Create the JWT AuthInfo instance
+        const authInfo = await AuthInfo.create({ username: testOrg.username });
+
+        // Verify the returned AuthInfo instance
+        const authInfoConnOpts = authInfo.getConnectionOptions();
+        expect(authInfoConnOpts).to.have.property('accessToken', testOrg.accessToken);
+        expect(authInfoConnOpts).to.have.property('instanceUrl', testOrg.instanceUrl);
+        expect(authInfoConnOpts).to.have.property('refreshFn').and.is.a('function');
+        expect(authInfo.getUsername()).to.equal(testOrg.username);
+        expect(authInfo.isAccessTokenFlow(), 'authInfo.isAccessTokenFlow() should be false').to.be.false;
+        expect(authInfo.isRefreshTokenFlow(), 'authInfo.isRefreshTokenFlow() should be false').to.be.false;
+        expect(authInfo.isJwt(), 'authInfo.isJwt() should be true').to.be.true;
+        expect(authInfo.isOauth(), 'authInfo.isOauth() should be false').to.be.false;
+
+        // Verify authInfo.fields are encrypted
+        expect(authInfo.getFields().accessToken).equals(jwtData.accessToken);
+      });
+
+      it('should throw an AuthInfoOverwriteError when both username and oauth data passed and auth file exists', async () => {
+        $$.setConfigStubContents('AuthInfoConfig', { contents: await testOrg.getConfig() });
+        stubMethod($$.SANDBOX, OrgAccessor.prototype, 'hasFile').resolves(true);
+
+        // Create the JWT AuthInfo instance
+        try {
+          await AuthInfo.create({
+            username: testOrg.username,
+            oauth2Options: {
+              clientId: testOrg.clientId,
+              loginUrl: testOrg.loginUrl,
+              privateKey: testOrg.privateKey,
+            },
+          });
+          assert.fail(
+            'Error thrown',
+            'No Error thrown',
+            'Expected AuthInfo.create() to throw an AuthInfoOverwriteError'
+          );
+        } catch (err) {
+          expect(err.name).to.equal('AuthInfoOverwriteError');
+        }
+      });
+
+      it('should throw a JWTAuthError when auth fails via a OAuth2.jwtAuthorize()', async () => {
+        const jwtConfig = {
+          clientId: testOrg.clientId,
+          loginUrl: testOrg.loginUrl,
+          privateKey: testOrg.privateKey,
+        };
+
+        // Stub file I/O, http requests, and the DNS lookup
+        stubMethod($$.SANDBOX, AuthInfo.prototype, 'readJwtKey').resolves('authInfoTest_private_key');
+        postParamsStub.throws(new Error('authInfoTest_ERROR_MSG'));
+        stubMethod($$.SANDBOX, jwt, 'sign').resolves('authInfoTest_jwtToken');
+        stubMethod($$.SANDBOX, dns, 'lookup').callsFake((url: string, done: (v: AnyJson, w: JsonMap) => {}) =>
+          done(null, { address: '1.1.1.1', family: 4 })
+        );
+
+        // Create the JWT AuthInfo instance
+        try {
+          await AuthInfo.create({ username: testOrg.username, oauth2Options: jwtConfig });
+          assert.fail('should have thrown an error within AuthInfo.authJwt()');
+        } catch (err) {
+          expect(err.name).to.equal('JwtAuthError');
+        }
+      });
+
+      it('should return a JWT AuthInfo instance when passed a username and JWT auth options despite failed DNS lookup', async () => {
+        $$.setConfigStubContents('AuthInfoConfig', { contents: await testOrg.getConfig() });
+
+        const jwtConfig = {
+          clientId: testOrg.clientId,
+          loginUrl: testOrg.loginUrl,
+          privateKey: testOrg.privateKey,
+        };
+        const jwtConfigClone = cloneJson(jwtConfig);
+        const authResponse = {
+          access_token: testOrg.accessToken,
+          instance_url: testOrg.instanceUrl,
+          id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
+        };
+
+        // Stub file I/O, http requests, and the DNS lookup
+        const readJwtKey = stubMethod($$.SANDBOX, AuthInfo.prototype, 'readJwtKey').resolves(
+          'authInfoTest_private_key'
+        );
+        postParamsStub.resolves(authResponse);
+        stubMethod($$.SANDBOX, jwt, 'sign').resolves('authInfoTest_jwtToken');
+        stubMethod($$.SANDBOX, SfdcUrl.prototype, 'lookup').throws();
+
+        // Create the JWT AuthInfo instance
+        const authInfo = await AuthInfo.create({
+          username: testOrg.username,
+          oauth2Options: jwtConfig,
+        });
+
+        // Verify the returned AuthInfo instance
+        const authInfoConnOpts = authInfo.getConnectionOptions();
+        expect(authInfoConnOpts).to.have.property('accessToken', authResponse.access_token);
+        expect(authInfoConnOpts).to.have.property('instanceUrl', testOrg.loginUrl);
+        expect(authInfoConnOpts).to.have.property('refreshFn').and.is.a('function');
+        expect(authInfo.getUsername()).to.equal(testOrg.username);
+        expect(authInfo.isAccessTokenFlow(), 'authInfo.isAccessTokenFlow() should be false').to.be.false;
+        expect(authInfo.isRefreshTokenFlow(), 'authInfo.isRefreshTokenFlow() should be false').to.be.false;
+        expect(authInfo.isJwt(), 'authInfo.isJwt() should be true').to.be.true;
+        expect(authInfo.isOauth(), 'authInfo.isOauth() should be false').to.be.false;
+
+        // Verify expected methods are called with expected args
+        expect(authInfoStubs.initAuthOptions.called).to.be.true;
+        expect(authInfoStubs.initAuthOptions.firstCall.args[0]).to.equal(jwtConfig);
+        expect(authInfoStubs.update.called).to.be.true;
+        expect(authInfoStubs.authJwt.called).to.be.true;
+        expect(authInfoStubs.authJwt.firstCall.args[0]).to.include(jwtConfig);
+        expect(
+          orgAccessorReadSpy.callCount,
+          'should have read an auth file once to ensure auth data did not already exist'
+        ).to.equal(1);
+        expect(readJwtKey.called).to.be.true;
+
+        // Verify the jwtConfig object was not mutated by init() or authJwt()
+        expect(jwtConfig).to.deep.equal(jwtConfigClone);
+
+        const expectedAuthConfig = {
+          accessToken: authResponse.access_token,
+          clientId: testOrg.clientId,
+          instanceUrl: testOrg.loginUrl,
+          orgId: authResponse.id.split('/')[0],
+          loginUrl: jwtConfig.loginUrl,
+          privateKey: jwtConfig.privateKey,
+          isDevHub: false,
+        };
+        expect(authInfoStubs.update.firstCall.args[0]).to.deep.equal(expectedAuthConfig);
       });
     });
 
-    it('should return a JWT AuthInfo instance when passed a username from an auth file', async () => {
-      const username = 'authInfoTest_username_jwt';
+    describe('Refresh Token', () => {
+      it('should return a refresh token AuthInfo instance when passed a username and refresh token auth options', async () => {
+        const refreshTokenConfig = {
+          refreshToken: testOrg.refreshToken,
+          loginUrl: testOrg.loginUrl,
+        };
+        const refreshTokenConfigClone = cloneJson(refreshTokenConfig);
+        const authResponse = {
+          access_token: testOrg.accessToken,
+          instance_url: testOrg.instanceUrl,
+          id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
+        };
 
-      // Make the file read stub return JWT auth data
-      const jwtData = {};
-      set(jwtData, 'accessToken', testMetadata.encryptedAccessToken);
-      set(jwtData, 'clientId', testMetadata.clientId);
-      set(jwtData, 'loginUrl', testMetadata.loginUrl);
-      set(jwtData, 'instanceUrl', testMetadata.instanceUrl);
-      set(jwtData, 'privateKey', 'authInfoTest/jwt/server.key');
-      set(jwtData, 'username', username);
-      testMetadata.fetchConfigInfo = () => jwtData;
+        // Stub the http request (OAuth2.refreshToken())
+        postParamsStub.resolves(authResponse);
 
-      // Create the JWT AuthInfo instance
-      const authInfo = await AuthInfo.create({ username });
+        // Create the refresh token AuthInfo instance
+        const authInfo = await AuthInfo.create({
+          username: testOrg.username,
+          oauth2Options: refreshTokenConfig,
+        });
 
-      // Verify the returned AuthInfo instance
-      const authInfoConnOpts = authInfo.getConnectionOptions();
-      expect(authInfoConnOpts).to.have.property('accessToken', testMetadata.accessToken);
-      expect(authInfoConnOpts).to.have.property('instanceUrl', testMetadata.instanceUrl);
-      expect(authInfoConnOpts).to.have.property('refreshFn').and.is.a('function');
-      expect(authInfo.getUsername()).to.equal(username);
-      expect(authInfo.isAccessTokenFlow(), 'authInfo.isAccessTokenFlow() should be false').to.be.false;
-      expect(authInfo.isRefreshTokenFlow(), 'authInfo.isRefreshTokenFlow() should be false').to.be.false;
-      expect(authInfo.isJwt(), 'authInfo.isJwt() should be true').to.be.true;
-      expect(authInfo.isOauth(), 'authInfo.isOauth() should be false').to.be.false;
+        verifyAuthInfoRefreshToken(authInfo, authResponse);
 
-      // Verify authInfo.fields are encrypted
-      expect(authInfo.getFields().accessToken).equals(getString(jwtData, 'accessToken'));
-    });
+        // Verify authInfo.fields are encrypted
+        const crypto = await Crypto.create();
+        expect(crypto.decrypt(authInfo.getFields().accessToken)).equals(authResponse.access_token);
+        expect(crypto.decrypt(authInfo.getFields().refreshToken)).equals(refreshTokenConfig.refreshToken);
 
-    it('should throw an AuthInfoOverwriteError when both username and oauth data passed and auth file exists', async () => {
-      const username = 'authInfoTest_username_jwt_from_auth_file';
-      const jwtConfig = {
-        clientId: testMetadata.clientId,
-        loginUrl: testMetadata.loginUrl,
-        privateKey: 'authInfoTest/jwt/server.key',
-      };
+        // Verify expected methods are called with expected args
+        expect(authInfoStubs.initAuthOptions.called).to.be.true;
+        expect(authInfoStubs.initAuthOptions.firstCall.args[0]).to.equal(refreshTokenConfig);
+        expect(authInfoStubs.update.called).to.be.true;
+        expect(authInfoStubs.buildRefreshTokenConfig.called).to.be.true;
+        expect(authInfoStubs.buildRefreshTokenConfig.firstCall.args[0]).to.include(refreshTokenConfig);
 
-      // Make the file read stub return JWT auth data
-      const jwtData = {};
-      set(jwtData, 'accessToken', testMetadata.encryptedAccessToken);
-      set(jwtData, 'clientId', testMetadata.clientId);
-      set(jwtData, 'loginUrl', testMetadata.loginUrl);
-      set(jwtData, 'instanceUrl', testMetadata.instanceUrl);
-      set(jwtData, 'privateKey', 'authInfoTest/jwt/server.key');
-      testMetadata.fetchConfigInfo = () => jwtData;
+        // Verify the refreshTokenConfig object was not mutated by init() or buildRefreshTokenConfig()
+        expect(refreshTokenConfig).to.deep.equal(refreshTokenConfigClone);
 
-      $$.setConfigStubContents('AuthInfoConfig', { contents: jwtData });
-      stubMethod($$.SANDBOX, OrgAccessor.prototype, 'hasFile').resolves(true);
-      // Create the JWT AuthInfo instance
-      try {
-        await AuthInfo.create({ username, oauth2Options: jwtConfig });
-        assert.fail('Error thrown', 'No Error thrown', 'Expected AuthInfo.create() to throw an AuthInfoOverwriteError');
-      } catch (err) {
-        expect(err.name).to.equal('AuthInfoOverwriteError');
-      }
-    });
-
-    it('should throw a JWTAuthError when auth fails via a OAuth2.jwtAuthorize()', async () => {
-      const username = 'authInfoTest_username_jwt_ERROR1';
-      const jwtConfig = {
-        clientId: testMetadata.clientId,
-        loginUrl: testMetadata.loginUrl,
-        privateKey: 'authInfoTest/jwt/server.key',
-      };
-
-      // Stub file I/O, http requests, and the DNS lookup
-      readFileStub.returns(Promise.resolve('authInfoTest_private_key'));
-      _postParmsStub.throws(new Error('authInfoTest_ERROR_MSG'));
-      stubMethod($$.SANDBOX, jwt, 'sign').returns(Promise.resolve('authInfoTest_jwtToken'));
-      stubMethod($$.SANDBOX, dns, 'lookup').callsFake((url: string, done: (v: AnyJson, w: JsonMap) => {}) =>
-        done(null, { address: '1.1.1.1', family: 4 })
-      );
-
-      // Create the JWT AuthInfo instance
-      try {
-        await AuthInfo.create({ username, oauth2Options: jwtConfig });
-        assert.fail('should have thrown an error within AuthInfo.authJwt()');
-      } catch (err) {
-        expect(err.name).to.equal('JwtAuthError');
-      }
-    });
-
-    //
-    // Refresh token tests
-    //
-
-    it('should return a refresh token AuthInfo instance when passed a username and refresh token auth options', async () => {
-      const username = 'authInfoTest_username_RefreshToken';
-      const refreshTokenConfig = {
-        refreshToken: testMetadata.refreshToken,
-        loginUrl: testMetadata.loginUrl,
-      };
-      const refreshTokenConfigClone = cloneJson(refreshTokenConfig);
-      const authResponse = {
-        access_token: testMetadata.accessToken,
-        instance_url: testMetadata.instanceUrl,
-        id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
-      };
-
-      // Stub the http request (OAuth2.refreshToken())
-      _postParmsStub.returns(Promise.resolve(authResponse));
-
-      // Create the refresh token AuthInfo instance
-      const authInfo = await AuthInfo.create({
-        username,
-        oauth2Options: refreshTokenConfig,
+        const expectedAuthConfig = {
+          accessToken: authResponse.access_token,
+          instanceUrl: testOrg.instanceUrl,
+          orgId: authResponse.id.split('/')[0],
+          loginUrl: refreshTokenConfig.loginUrl,
+          refreshToken: refreshTokenConfig.refreshToken,
+          clientId: testOrg.defaultConnectedAppInfo.clientId,
+          clientSecret: testOrg.defaultConnectedAppInfo.clientSecret,
+          isDevHub: false,
+          username: testOrg.username,
+        };
+        expect(authInfoStubs.update.firstCall.args[0]).to.deep.equal(expectedAuthConfig);
       });
 
-      // Verify the returned AuthInfo instance
-      const authInfoConnOpts = authInfo.getConnectionOptions();
-      expect(authInfoConnOpts).to.have.property('accessToken', authResponse.access_token);
-      expect(authInfoConnOpts).to.have.property('instanceUrl', authResponse.instance_url);
-      expect(authInfoConnOpts).to.not.have.property('refreshToken');
-      expect(authInfoConnOpts['oauth2']).to.have.property('loginUrl', testMetadata.instanceUrl);
-      expect(authInfoConnOpts['oauth2']).to.have.property('clientId', testMetadata.defaultConnectedAppInfo.clientId);
-      expect(authInfoConnOpts['oauth2']).to.have.property('redirectUri', testMetadata.redirectUri);
-      expect(authInfo.getUsername()).to.equal(username);
-      expect(authInfo.isAccessTokenFlow(), 'authInfo.isAccessTokenFlow() should be false').to.be.false;
-      expect(authInfo.isRefreshTokenFlow(), 'authInfo.isRefreshTokenFlow() should be true').to.be.true;
-      expect(authInfo.isJwt(), 'authInfo.isJwt() should be false').to.be.false;
-      expect(authInfo.isOauth(), 'authInfo.isOauth() should be true').to.be.true;
+      it('should return a refresh token AuthInfo instance with username in auth options', async () => {
+        const refreshTokenConfig = {
+          refreshToken: testOrg.refreshToken,
+          loginUrl: testOrg.loginUrl,
+          username: testOrg.username,
+        };
+        const refreshTokenConfigClone = cloneJson(refreshTokenConfig);
+        const authResponse = {
+          access_token: testOrg.accessToken,
+          instance_url: testOrg.instanceUrl,
+          id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
+        };
 
-      // Verify authInfo.fields are encrypted
-      const crypto = await Crypto.create();
-      expect(crypto.decrypt(authInfo.getFields().accessToken)).equals(authResponse.access_token);
-      expect(crypto.decrypt(authInfo.getFields().refreshToken)).equals(refreshTokenConfig.refreshToken);
+        // Stub the http request (OAuth2.refreshToken())
+        postParamsStub.resolves(authResponse);
 
-      // Verify expected methods are called with expected args
-      expect(authInfoInit.called).to.be.true;
-      expect(authInfoInit.firstCall.args[0]).to.equal(refreshTokenConfig);
-      expect(authInfoUpdate.called).to.be.true;
-      expect(authInfoBuildRefreshTokenConfig.called).to.be.true;
-      expect(authInfoBuildRefreshTokenConfig.firstCall.args[0]).to.include(refreshTokenConfig);
+        // Create the refresh token AuthInfo instance
+        const authInfo = await AuthInfo.create({
+          oauth2Options: refreshTokenConfig,
+        });
 
-      // Verify the refreshTokenConfig object was not mutated by init() or buildRefreshTokenConfig()
-      expect(refreshTokenConfig).to.deep.equal(refreshTokenConfigClone);
+        verifyAuthInfoRefreshToken(authInfo, authResponse);
 
-      const expectedAuthConfig = {
-        accessToken: authResponse.access_token,
-        instanceUrl: testMetadata.instanceUrl,
-        orgId: authResponse.id.split('/')[0],
-        loginUrl: refreshTokenConfig.loginUrl,
-        refreshToken: refreshTokenConfig.refreshToken,
-        clientId: testMetadata.defaultConnectedAppInfo.clientId,
-        clientSecret: testMetadata.defaultConnectedAppInfo.clientSecret,
-        isDevHub: false,
-        username,
-      };
-      expect(authInfoUpdate.firstCall.args[0]).to.deep.equal(expectedAuthConfig);
-    });
+        // Verify authInfo.fields are encrypted
+        const crypto = await Crypto.create();
+        expect(crypto.decrypt(authInfo.getFields().accessToken)).equals(authResponse.access_token);
+        expect(crypto.decrypt(authInfo.getFields().refreshToken)).equals(refreshTokenConfig.refreshToken);
 
-    it('should return a refresh token AuthInfo instance with username in auth options', async () => {
-      const username = 'authInfoTest_username_RefreshToken';
-      const refreshTokenConfig = {
-        refreshToken: testMetadata.refreshToken,
-        loginUrl: testMetadata.loginUrl,
-        username,
-      };
-      const refreshTokenConfigClone = cloneJson(refreshTokenConfig);
-      const authResponse = {
-        access_token: testMetadata.accessToken,
-        instance_url: testMetadata.instanceUrl,
-        id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
-      };
+        // Verify expected methods are called with expected args
+        expect(authInfoStubs.initAuthOptions.called).to.be.true;
+        expect(authInfoStubs.initAuthOptions.firstCall.args[0]).to.equal(refreshTokenConfig);
+        expect(authInfoStubs.update.called).to.be.true;
+        expect(authInfoStubs.buildRefreshTokenConfig.called).to.be.true;
+        expect(authInfoStubs.buildRefreshTokenConfig.firstCall.args[0]).to.include(refreshTokenConfig);
 
-      // Stub the http request (OAuth2.refreshToken())
-      _postParmsStub.returns(Promise.resolve(authResponse));
+        // Verify the refreshTokenConfig object was not mutated by init() or buildRefreshTokenConfig()
+        expect(refreshTokenConfig).to.deep.equal(refreshTokenConfigClone);
 
-      // Create the refresh token AuthInfo instance
-      const authInfo = await AuthInfo.create({
-        oauth2Options: refreshTokenConfig,
+        const expectedAuthConfig = {
+          accessToken: authResponse.access_token,
+          instanceUrl: testOrg.instanceUrl,
+          orgId: authResponse.id.split('/')[0],
+          loginUrl: refreshTokenConfig.loginUrl,
+          refreshToken: refreshTokenConfig.refreshToken,
+          clientId: testOrg.defaultConnectedAppInfo.clientId,
+          clientSecret: testOrg.defaultConnectedAppInfo.clientSecret,
+          isDevHub: false,
+          username: testOrg.username,
+        };
+        expect(authInfoStubs.update.firstCall.args[0]).to.deep.equal(expectedAuthConfig);
       });
 
-      // Verify the returned AuthInfo instance
-      const authInfoConnOpts = authInfo.getConnectionOptions();
-      expect(authInfoConnOpts).to.have.property('accessToken', authResponse.access_token);
-      expect(authInfoConnOpts).to.have.property('instanceUrl', authResponse.instance_url);
-      expect(authInfoConnOpts).to.not.have.property('refreshToken');
-      expect(authInfoConnOpts['oauth2']).to.have.property('loginUrl', testMetadata.instanceUrl);
-      expect(authInfoConnOpts['oauth2']).to.have.property('clientId', testMetadata.defaultConnectedAppInfo.clientId);
-      expect(authInfoConnOpts['oauth2']).to.have.property('redirectUri', testMetadata.redirectUri);
-      expect(authInfo.getUsername()).to.equal(username);
-      expect(authInfo.isAccessTokenFlow(), 'authInfo.isAccessTokenFlow() should be false').to.be.false;
-      expect(authInfo.isRefreshTokenFlow(), 'authInfo.isRefreshTokenFlow() should be true').to.be.true;
-      expect(authInfo.isJwt(), 'authInfo.isJwt() should be false').to.be.false;
-      expect(authInfo.isOauth(), 'authInfo.isOauth() should be true').to.be.true;
+      it('should return a refresh token AuthInfo instance without any username', async () => {
+        const refreshTokenConfig = {
+          refreshToken: testOrg.refreshToken,
+          loginUrl: testOrg.loginUrl,
+        };
+        const refreshTokenConfigClone = cloneJson(refreshTokenConfig);
+        const authResponse = {
+          access_token: testOrg.accessToken,
+          instance_url: testOrg.instanceUrl,
+          id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
+        };
 
-      // Verify authInfo.fields are encrypted
-      const crypto = await Crypto.create();
-      expect(crypto.decrypt(authInfo.getFields().accessToken)).equals(authResponse.access_token);
-      expect(crypto.decrypt(authInfo.getFields().refreshToken)).equals(refreshTokenConfig.refreshToken);
+        // Stub the http request (OAuth2.refreshToken())
+        postParamsStub.resolves(authResponse);
+        stubUserRequest();
 
-      // Verify expected methods are called with expected args
-      expect(authInfoInit.called).to.be.true;
-      expect(authInfoInit.firstCall.args[0]).to.equal(refreshTokenConfig);
-      expect(authInfoUpdate.called).to.be.true;
-      expect(authInfoBuildRefreshTokenConfig.called).to.be.true;
-      expect(authInfoBuildRefreshTokenConfig.firstCall.args[0]).to.include(refreshTokenConfig);
+        // Create the refresh token AuthInfo instance
+        const authInfo = await AuthInfo.create({
+          oauth2Options: refreshTokenConfig,
+        });
 
-      // Verify the refreshTokenConfig object was not mutated by init() or buildRefreshTokenConfig()
-      expect(refreshTokenConfig).to.deep.equal(refreshTokenConfigClone);
+        // Verify the returned AuthInfo instance
+        const authInfoConnOpts = authInfo.getConnectionOptions();
+        expect(authInfoConnOpts).to.have.property('accessToken', authResponse.access_token);
+        expect(authInfoConnOpts).to.have.property('instanceUrl', authResponse.instance_url);
+        expect(authInfoConnOpts).to.not.have.property('refreshToken');
+        expect(authInfoConnOpts['oauth2']).to.have.property('loginUrl', testOrg.instanceUrl);
+        expect(authInfoConnOpts['oauth2']).to.have.property('clientId', testOrg.defaultConnectedAppInfo.clientId);
+        expect(authInfoConnOpts['oauth2']).to.have.property('redirectUri', testOrg.redirectUri);
+        expect(authInfo.getUsername()).to.equal(testOrg.username.toUpperCase());
+        expect(authInfo.isAccessTokenFlow(), 'authInfo.isAccessTokenFlow() should be false').to.be.false;
+        expect(authInfo.isRefreshTokenFlow(), 'authInfo.isRefreshTokenFlow() should be true').to.be.true;
+        expect(authInfo.isJwt(), 'authInfo.isJwt() should be false').to.be.false;
+        expect(authInfo.isOauth(), 'authInfo.isOauth() should be true').to.be.true;
 
-      const expectedAuthConfig = {
-        accessToken: authResponse.access_token,
-        instanceUrl: testMetadata.instanceUrl,
-        orgId: authResponse.id.split('/')[0],
-        loginUrl: refreshTokenConfig.loginUrl,
-        refreshToken: refreshTokenConfig.refreshToken,
-        clientId: testMetadata.defaultConnectedAppInfo.clientId,
-        clientSecret: testMetadata.defaultConnectedAppInfo.clientSecret,
-        isDevHub: false,
-        username,
-      };
-      expect(authInfoUpdate.firstCall.args[0]).to.deep.equal(expectedAuthConfig);
-    });
+        // Verify authInfo.fields are encrypted
+        const crypto = await Crypto.create();
+        expect(crypto.decrypt(authInfo.getFields().accessToken)).equals(authResponse.access_token);
+        expect(crypto.decrypt(authInfo.getFields().refreshToken)).equals(refreshTokenConfig.refreshToken);
 
-    it('should return a refresh token AuthInfo instance with custom clientId and clientSecret', async () => {
-      const username = 'authInfoTest_username_RefreshToken_Custom';
-      const refreshTokenConfig = {
-        clientId: 'authInfoTest_clientId',
-        clientSecret: 'authInfoTest_clientSecret',
-        refreshToken: testMetadata.refreshToken,
-        loginUrl: testMetadata.loginUrl,
-        redirectUri: 'http://localhost:1717/OauthRedirect',
-      };
-      const authResponse = {
-        access_token: testMetadata.accessToken,
-        instance_url: testMetadata.instanceUrl,
-        id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
-      };
+        // Verify expected methods are called with expected args
+        expect(authInfoStubs.initAuthOptions.called).to.be.true;
+        expect(authInfoStubs.initAuthOptions.firstCall.args[0]).to.equal(refreshTokenConfig);
+        expect(authInfoStubs.update.called).to.be.true;
+        expect(authInfoStubs.buildRefreshTokenConfig.called).to.be.true;
+        expect(authInfoStubs.buildRefreshTokenConfig.firstCall.args[0]).to.include(refreshTokenConfig);
 
-      // Stub the http request (OAuth2.refreshToken())
-      _postParmsStub.returns(Promise.resolve(authResponse));
+        // Verify the refreshTokenConfig object was not mutated by init() or buildRefreshTokenConfig()
+        expect(refreshTokenConfig).to.deep.equal(refreshTokenConfigClone);
 
-      // Create the refresh token AuthInfo instance
-      const authInfo = await AuthInfo.create({
-        username,
-        oauth2Options: refreshTokenConfig,
+        const expectedAuthConfig = {
+          accessToken: authResponse.access_token,
+          instanceUrl: testOrg.instanceUrl,
+          orgId: authResponse.id.split('/')[0],
+          loginUrl: refreshTokenConfig.loginUrl,
+          refreshToken: refreshTokenConfig.refreshToken,
+          clientId: testOrg.defaultConnectedAppInfo.clientId,
+          clientSecret: testOrg.defaultConnectedAppInfo.clientSecret,
+          isDevHub: false,
+          username: testOrg.username.toUpperCase(),
+        };
+        expect(authInfoStubs.update.firstCall.args[0]).to.deep.equal(expectedAuthConfig);
       });
 
-      // Verify the returned AuthInfo instance
-      const authInfoConnOpts = authInfo.getConnectionOptions();
-      expect(authInfoConnOpts).to.have.property('accessToken', authResponse.access_token);
-      expect(authInfoConnOpts).to.have.property('instanceUrl', authResponse.instance_url);
-      expect(authInfoConnOpts).to.not.have.property('refreshToken');
-      expect(authInfoConnOpts['oauth2']).to.have.property('loginUrl', testMetadata.instanceUrl);
-      expect(authInfoConnOpts['oauth2']).to.have.property('clientId', refreshTokenConfig.clientId);
-      expect(authInfoConnOpts['oauth2']).to.have.property('redirectUri', testMetadata.redirectUri);
-      expect(authInfo.getUsername()).to.equal(username);
-      expect(authInfo.isAccessTokenFlow(), 'authInfo.isAccessTokenFlow() should be false').to.be.false;
-      expect(authInfo.isRefreshTokenFlow(), 'authInfo.isRefreshTokenFlow() should be true').to.be.true;
-      expect(authInfo.isJwt(), 'authInfo.isJwt() should be false').to.be.false;
-      expect(authInfo.isOauth(), 'authInfo.isOauth() should be true').to.be.true;
+      it('should return a refresh token AuthInfo instance with custom clientId and clientSecret', async () => {
+        const refreshTokenConfig = {
+          clientId: 'authInfoTest_clientId',
+          clientSecret: 'authInfoTest_clientSecret',
+          refreshToken: testOrg.refreshToken,
+          loginUrl: testOrg.loginUrl,
+          redirectUri: 'http://localhost:1717/OauthRedirect',
+        };
+        const authResponse = {
+          access_token: testOrg.accessToken,
+          instance_url: testOrg.instanceUrl,
+          id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
+        };
 
-      // Verify authInfo.fields are encrypted
-      const crypto = await Crypto.create();
-      expect(crypto.decrypt(authInfo.getFields().accessToken)).equals(authResponse.access_token);
-      expect(crypto.decrypt(authInfo.getFields().refreshToken)).equals(refreshTokenConfig.refreshToken);
-      expect(crypto.decrypt(authInfo.getFields().clientSecret)).equals(refreshTokenConfig.clientSecret);
+        // Stub the http request (OAuth2.refreshToken())
+        postParamsStub.resolves(authResponse);
 
-      // Verify expected methods are called with expected args
-      expect(authInfoInit.called).to.be.true;
-      expect(authInfoInit.firstCall.args[0]).to.equal(refreshTokenConfig);
-      expect(authInfoUpdate.called).to.be.true;
-      expect(authInfoBuildRefreshTokenConfig.called).to.be.true;
-      expect(authInfoBuildRefreshTokenConfig.firstCall.args[0]).to.deep.equal(refreshTokenConfig);
+        // Create the refresh token AuthInfo instance
+        const authInfo = await AuthInfo.create({
+          username: testOrg.username,
+          oauth2Options: refreshTokenConfig,
+        });
 
-      const expectedAuthConfig = {
-        accessToken: authResponse.access_token,
-        instanceUrl: testMetadata.instanceUrl,
-        orgId: authResponse.id.split('/')[0],
-        loginUrl: refreshTokenConfig.loginUrl,
-        refreshToken: refreshTokenConfig.refreshToken,
-        clientId: refreshTokenConfig.clientId,
-        clientSecret: refreshTokenConfig.clientSecret,
-        isDevHub: false,
-        username,
-      };
-      expect(authInfoUpdate.firstCall.args[0]).to.deep.equal(expectedAuthConfig);
+        verifyAuthInfoRefreshToken(authInfo, authResponse, refreshTokenConfig);
+
+        // Verify authInfo.fields are encrypted
+        const crypto = await Crypto.create();
+        expect(crypto.decrypt(authInfo.getFields().accessToken)).equals(authResponse.access_token);
+        expect(crypto.decrypt(authInfo.getFields().refreshToken)).equals(refreshTokenConfig.refreshToken);
+        expect(crypto.decrypt(authInfo.getFields().clientSecret)).equals(refreshTokenConfig.clientSecret);
+
+        // Verify expected methods are called with expected args
+        expect(authInfoStubs.initAuthOptions.called).to.be.true;
+        expect(authInfoStubs.initAuthOptions.firstCall.args[0]).to.equal(refreshTokenConfig);
+        expect(authInfoStubs.update.called).to.be.true;
+        expect(authInfoStubs.buildRefreshTokenConfig.called).to.be.true;
+        expect(authInfoStubs.buildRefreshTokenConfig.firstCall.args[0]).to.include(refreshTokenConfig);
+
+        const expectedAuthConfig = {
+          accessToken: authResponse.access_token,
+          instanceUrl: testOrg.instanceUrl,
+          orgId: authResponse.id.split('/')[0],
+          loginUrl: refreshTokenConfig.loginUrl,
+          refreshToken: refreshTokenConfig.refreshToken,
+          clientId: refreshTokenConfig.clientId,
+          clientSecret: refreshTokenConfig.clientSecret,
+          isDevHub: false,
+          username: testOrg.username,
+        };
+        expect(authInfoStubs.update.firstCall.args[0]).to.deep.equal(expectedAuthConfig);
+      });
+
+      it('should throw a RefreshTokenAuthError when auth fails via a refresh token', async () => {
+        const username = 'authInfoTest_username_RefreshToken_ERROR';
+        const refreshTokenConfig = {
+          clientId: 'authInfoTest_clientId',
+          clientSecret: 'authInfoTest_clientSecret',
+          refreshToken: testOrg.refreshToken,
+          loginUrl: testOrg.loginUrl,
+        };
+
+        // Stub the http request (OAuth2.refreshToken())
+        postParamsStub.throws(new Error('authInfoTest_ERROR_MSG'));
+
+        // Create the refresh token AuthInfo instance
+        try {
+          await AuthInfo.create({ username, oauth2Options: refreshTokenConfig });
+          assert.fail('should have thrown an error within AuthInfo.buildRefreshTokenConfig()');
+        } catch (err) {
+          expect(err.name).to.equal('RefreshTokenAuthError');
+        }
+      });
     });
 
-    it('should throw a RefreshTokenAuthError when auth fails via a refresh token', async () => {
-      const username = 'authInfoTest_username_RefreshToken_ERROR';
-      const refreshTokenConfig = {
-        clientId: 'authInfoTest_clientId',
-        clientSecret: 'authInfoTest_clientSecret',
-        refreshToken: testMetadata.refreshToken,
-        loginUrl: testMetadata.loginUrl,
-      };
+    describe('Web Auth', () => {
+      it('should return a refresh token AuthInfo instance when passed an authcode', async () => {
+        const authCodeConfig = {
+          authCode: testOrg.authcode,
+          loginUrl: testOrg.loginUrl,
+        };
+        const authCodeConfigClone = cloneJson(authCodeConfig);
+        const authResponse = {
+          access_token: testOrg.accessToken,
+          instance_url: testOrg.instanceUrl,
+          id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
+          refresh_token: testOrg.refreshToken,
+        };
 
-      // Stub the http request (OAuth2.refreshToken())
-      _postParmsStub.throws(new Error('authInfoTest_ERROR_MSG'));
+        // Stub the http requests (OAuth2.requestToken() and the request for the username)
+        postParamsStub.resolves(authResponse);
+        const stub = stubUserRequest();
 
-      // Create the refresh token AuthInfo instance
-      try {
-        await AuthInfo.create({ username, oauth2Options: refreshTokenConfig });
-        assert.fail('should have thrown an error within AuthInfo.buildRefreshTokenConfig()');
-      } catch (err) {
-        expect(err.name).to.equal('RefreshTokenAuthError');
-      }
-    });
+        // Create the refresh token AuthInfo instance
+        const authInfo = await AuthInfo.create({ oauth2Options: authCodeConfig });
 
-    //
-    // Web Auth (auth code) tests
-    //
+        // Ensure we query for the username
+        expect(stub.called).to.be.true;
+        const authInfoConnOpts = authInfo.getConnectionOptions();
+        expect(authInfoConnOpts).to.have.property('accessToken', authResponse.access_token);
+        expect(authInfoConnOpts).to.have.property('instanceUrl', authResponse.instance_url);
+        expect(authInfoConnOpts).to.not.have.property('refreshToken');
+        expect(authInfoConnOpts['oauth2']).to.have.property('loginUrl', testOrg.instanceUrl); // why is this instanceUrl?
+        expect(authInfoConnOpts['oauth2']).to.have.property('clientId', testOrg.defaultConnectedAppInfo.clientId);
+        expect(authInfoConnOpts['oauth2']).to.have.property('redirectUri', testOrg.redirectUri);
+        expect(authInfo.getUsername()).to.equal(testOrg.username.toUpperCase());
+        expect(authInfo.isAccessTokenFlow(), 'authInfo.isAccessTokenFlow() should be false').to.be.false;
+        expect(authInfo.isRefreshTokenFlow(), 'authInfo.isRefreshTokenFlow() should be true').to.be.true;
+        expect(authInfo.isJwt(), 'authInfo.isJwt() should be false').to.be.false;
+        expect(authInfo.isOauth(), 'authInfo.isOauth() should be true').to.be.true;
 
-    it('should return a refresh token AuthInfo instance when passed an authcode', async () => {
-      const username = 'authInfoTest_username_AuthCode';
-      const authCodeConfig = {
-        authCode: testMetadata.authCode,
-        loginUrl: testMetadata.loginUrl,
-      };
-      const authCodeConfigClone = cloneJson(authCodeConfig);
-      const authResponse = {
-        access_token: testMetadata.accessToken,
-        instance_url: testMetadata.instanceUrl,
-        id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
-        refresh_token: testMetadata.refreshToken,
-      };
+        // Verify authInfo.fields are encrypted
+        const crypto = await Crypto.create();
+        expect(crypto.decrypt(authInfo.getFields().accessToken)).equals(authResponse.access_token);
+        expect(crypto.decrypt(authInfo.getFields().refreshToken)).equals(authResponse.refresh_token);
 
-      // Stub the http requests (OAuth2.requestToken() and the request for the username)
-      _postParmsStub.returns(Promise.resolve(authResponse));
-      const userInfoResponseBody = {
-        body: JSON.stringify({ preferred_username: username, organization_id: testMetadata.orgId }),
-      };
-      const userResponseBody = {
-        body: JSON.stringify({ Username: username.toUpperCase() }),
-      };
-      const stub = stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest')
-        .onFirstCall()
-        .returns(Promise.resolve(userInfoResponseBody))
-        .onSecondCall()
-        .returns(Promise.resolve(userResponseBody))
-        .onThirdCall()
-        .throws();
+        // Verify expected methods are called with expected args
+        expect(authInfoStubs.initAuthOptions.called).to.be.true;
+        expect(authInfoStubs.initAuthOptions.firstCall.args[0]).to.equal(authCodeConfig);
+        expect(authInfoStubs.update.called).to.be.true;
+        expect(authInfoStubs.exchangeToken.called).to.be.true;
+        expect(authInfoStubs.exchangeToken.firstCall.args[0]).to.include(authCodeConfig);
 
-      // Create the refresh token AuthInfo instance
-      const authInfo = await AuthInfo.create({ oauth2Options: authCodeConfig });
+        // Verify the authCodeConfig object was not mutated by init() or buildWebAuthConfig()
+        expect(authCodeConfig).to.deep.equal(authCodeConfigClone);
 
-      // Ensure we query for the username
-      expect(stub.called).to.be.true;
+        const expectedAuthConfig = {
+          accessToken: authResponse.access_token,
+          instanceUrl: testOrg.instanceUrl,
+          username: testOrg.username.toUpperCase(),
+          orgId: authResponse.id.split('/')[0],
+          loginUrl: authCodeConfig.loginUrl,
+          refreshToken: authResponse.refresh_token,
+          isDevHub: false,
+          // These need to be passed in by the consumer. Since they are not, they will show up as undefined.
+          // In a non-test environment, the exchange will fail because no clientId is supplied.
+          clientId: undefined,
+          clientSecret: undefined,
+        };
+        expect(authInfoStubs.update.firstCall.args[0]).to.deep.equal(expectedAuthConfig);
+      });
 
-      // Verify the returned AuthInfo instance
-      const authInfoConnOpts = authInfo.getConnectionOptions();
-      expect(authInfoConnOpts).to.have.property('accessToken', authResponse.access_token);
-      expect(authInfoConnOpts).to.have.property('instanceUrl', authResponse.instance_url);
-      expect(authInfoConnOpts).to.not.have.property('refreshToken');
-      expect(authInfoConnOpts['oauth2']).to.have.property('loginUrl', testMetadata.instanceUrl); // why is this instanceUrl?
-      expect(authInfoConnOpts['oauth2']).to.have.property('clientId', testMetadata.defaultConnectedAppInfo.clientId);
-      expect(authInfoConnOpts['oauth2']).to.have.property('redirectUri', testMetadata.redirectUri);
-      expect(authInfo.getUsername()).to.equal(username.toUpperCase());
-      expect(authInfo.isAccessTokenFlow(), 'authInfo.isAccessTokenFlow() should be false').to.be.false;
-      expect(authInfo.isRefreshTokenFlow(), 'authInfo.isRefreshTokenFlow() should be true').to.be.true;
-      expect(authInfo.isJwt(), 'authInfo.isJwt() should be false').to.be.false;
-      expect(authInfo.isOauth(), 'authInfo.isOauth() should be true').to.be.true;
+      it('should return access token and refresh token when using authCode; verifier should be the same.', async () => {
+        /**
+         * The way web oauth works is first you must request a one-time auth code. Typically
+         * you send a hashed number to the server; the number is called the code verifier and the hashed value the code_challenge
+         * After the code is returned you then request an access token by passing along the returned code
+         * an you also include the UNHASHED code verifier value. If these two items match from when the authcode was issued the
+         * access/refresh tokens are then returned.
+         *
+         * Typically the authCode is obtained by authenticating to salesforce via a generated url that contains the
+         * connected app info plus this code challenge. After successful authentication the browser is sent a redirect url
+         * that includes the authCode.
+         *
+         * This test just makes sure the auth code exchange method is using the same codeVerifier. By creating the
+         * codeVerifier instance you can first generate the url then pass it to AuthInfo. Then everything lines up.
+         */
+        const options: OAuth2Config & { authCode?: string } = {
+          clientId: testOrg.clientId,
+          clientSecret: testOrg.clientSecret,
+          loginUrl: testOrg.loginUrl,
+          redirectUri: testOrg.redirectUri,
+        };
+        const oauth2 = new OAuth2(options);
+        options.authCode = '123456';
 
-      // Verify authInfo.fields are encrypted
-      const crypto = await Crypto.create();
-      expect(crypto.decrypt(authInfo.getFields().accessToken)).equals(authResponse.access_token);
-      expect(crypto.decrypt(authInfo.getFields().refreshToken)).equals(authResponse.refresh_token);
+        const authResponse = {
+          access_token: testOrg.accessToken,
+          instance_url: testOrg.instanceUrl,
+          id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
+          refresh_token: testOrg.refreshToken,
+        };
+        // Stub the http requests (OAuth2.requestToken() and the request for the username)
+        postParamsStub.resolves(authResponse);
+        stubUserRequest();
 
-      // Verify expected methods are called with expected args
-      expect(authInfoInit.called).to.be.true;
-      expect(authInfoInit.firstCall.args[0]).to.equal(authCodeConfig);
-      expect(authInfoUpdate.called).to.be.true;
-      expect(authInfoExchangeToken.called).to.be.true;
-      expect(authInfoExchangeToken.firstCall.args[0]).to.include(authCodeConfig);
+        await AuthInfo.create({ oauth2Options: options, oauth2 });
+        expect(authInfoStubs.exchangeToken.args.length).to.equal(1);
+        expect(authInfoStubs.exchangeToken.args[0].length).to.equal(2);
+        expect(authInfoStubs.exchangeToken.args[0][1]).to.have.property('codeVerifier', oauth2.codeVerifier);
+      });
 
-      // Verify the authCodeConfig object was not mutated by init() or buildWebAuthConfig()
-      expect(authCodeConfig).to.deep.equal(authCodeConfigClone);
+      it('should throw a AuthCodeExchangeError when auth fails via an auth code', async () => {
+        const authCodeConfig = {
+          authCode: testOrg.authcode,
+          loginUrl: testOrg.loginUrl,
+        };
 
-      const expectedAuthConfig = {
-        accessToken: authResponse.access_token,
-        instanceUrl: testMetadata.instanceUrl,
-        username: username.toUpperCase(),
-        orgId: authResponse.id.split('/')[0],
-        loginUrl: authCodeConfig.loginUrl,
-        refreshToken: authResponse.refresh_token,
-        isDevHub: false,
-        // These need to be passed in by the consumer. Since they are not, they will show up as undefined.
-        // In a non-test environment, the exchange will fail because no clientId is supplied.
-        clientId: undefined,
-        clientSecret: undefined,
-      };
-      expect(authInfoUpdate.firstCall.args[0]).to.deep.equal(expectedAuthConfig);
-    });
+        // Stub the http request (OAuth2.requestToken())
+        postParamsStub.throws(new Error('authInfoTest_ERROR_MSG'));
 
-    it('should return access token and refresh token when using authCode; verifier should be the same.', async () => {
-      /**
-       * The way web oauth works is first you must request a one-time auth code. Typically
-       * you send a hashed number to the server; the number is called the code verifier and the hashed value the code_challenge
-       * After the code is returned you then request an access token by passing along the returned code
-       * an you also include the UNHASHED code verifier value. If these two items match from when the authcode was issued the
-       * access/refresh tokens are then returned.
-       *
-       * Typically the authCode is obtained by authenticating to salesforce via a generated url that contains the
-       * connected app info plus this code challenge. After successful authentication the browser is sent a redirect url
-       * that includes the authCode.
-       *
-       * This test just makes sure the auth code exchange method is using the same codeVerifier. By creating the
-       * codeVerifier instance you can first generate the url then pass it to AuthInfo. Then everything lines up.
-       */
-      const clientId = 'clientId';
-      const clientSecret = 'clientSecret';
-      const loginUrl = 'loginUrl';
-      const redirectUri = 'redirectUri';
-      const username = 'authInfoTest_username_AuthCode';
+        // Create the auth code AuthInfo instance
+        try {
+          await AuthInfo.create({ oauth2Options: authCodeConfig });
+          assert.fail('should have thrown an error within AuthInfo.buildWebAuthConfig()');
+        } catch (err) {
+          expect(err.name).to.equal('AuthCodeExchangeError');
+        }
+      });
 
-      const options: OAuth2Config & { authCode?: string } = { clientId, clientSecret, loginUrl, redirectUri };
-      const oauth2 = new OAuth2(options);
-      options.authCode = '123456';
+      it('should throw a AuthCodeUsernameRetrievalError when userInfo retrieval fails after auth code exchange', async () => {
+        const authCodeConfig = {
+          authCode: testOrg.authcode,
+          loginUrl: testOrg.loginUrl,
+        };
+        const authResponse = {
+          access_token: testOrg.accessToken,
+          instance_url: testOrg.instanceUrl,
+          id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
+          refresh_token: testOrg.refreshToken,
+        };
 
-      const authResponse = {
-        access_token: testMetadata.accessToken,
-        instance_url: testMetadata.instanceUrl,
-        id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
-        refresh_token: testMetadata.refreshToken,
-      };
-      // Stub the http requests (OAuth2.requestToken() and the request for the username)
-      _postParmsStub.returns(Promise.resolve(authResponse));
-      const userInfoResponseBody = {
-        body: JSON.stringify({ preferred_username: username, organization_id: testMetadata.orgId }),
-      };
-      const userResponseBody = {
-        body: JSON.stringify({ Username: username.toUpperCase() }),
-      };
-      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest')
-        .onFirstCall()
-        .returns(Promise.resolve(userInfoResponseBody))
-        .onSecondCall()
-        .returns(Promise.resolve(userResponseBody));
-      await AuthInfo.create({ oauth2Options: options, oauth2 });
-      expect(authInfoExchangeToken.args.length).to.equal(1);
-      expect(authInfoExchangeToken.args[0].length).to.equal(2);
-      expect(authInfoExchangeToken.args[0][1]).to.have.property('codeVerifier', oauth2.codeVerifier);
-    });
+        // Stub the http request (OAuth2.requestToken())
+        postParamsStub.resolves(authResponse);
+        stubUserRequest({
+          statusCode: 404,
+          body: [
+            {
+              message: 'Could not retrieve the username after successful auth code exchange.\nDue to: %s',
+              errorCode: 'AuthCodeUsernameRetrievalError',
+            },
+          ],
+        });
 
-    it('should throw a AuthCodeExchangeError when auth fails via an auth code', async () => {
-      const authCodeConfig = {
-        authCode: testMetadata.authCode,
-        loginUrl: testMetadata.loginUrl,
-      };
+        // Create the auth code AuthInfo instance
+        try {
+          await AuthInfo.create({ oauth2Options: authCodeConfig });
+          assert.fail('should have thrown an error within AuthInfo.buildWebAuthConfig()');
+        } catch (err) {
+          expect(err.name).to.equal('AuthCodeUsernameRetrievalError');
+        }
+      });
 
-      // Stub the http request (OAuth2.requestToken())
-      _postParmsStub.throws(new Error('authInfoTest_ERROR_MSG'));
+      it('should throw a AuthCodeUsernameRetrievalError when user sobject retrieval fails after auth code exchange', async () => {
+        const authCodeConfig = {
+          authCode: testOrg.authcode,
+          loginUrl: testOrg.loginUrl,
+        };
+        const authResponse = {
+          access_token: testOrg.accessToken,
+          instance_url: testOrg.instanceUrl,
+          id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
+          refresh_token: testOrg.refreshToken,
+        };
 
-      // Create the auth code AuthInfo instance
-      try {
-        await AuthInfo.create({ oauth2Options: authCodeConfig });
-        assert.fail('should have thrown an error within AuthInfo.buildWebAuthConfig()');
-      } catch (err) {
-        expect(err.name).to.equal('AuthCodeExchangeError');
-      }
-    });
+        // Stub the http request (OAuth2.requestToken())
+        postParamsStub.resolves(authResponse);
 
-    it('should throw a AuthCodeUsernameRetrievalError when userInfo retrieval fails after auth code exchange', async () => {
-      const authCodeConfig = {
-        authCode: testMetadata.authCode,
-        loginUrl: testMetadata.loginUrl,
-      };
-      const authResponse = {
-        access_token: testMetadata.accessToken,
-        instance_url: testMetadata.instanceUrl,
-        id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
-        refresh_token: testMetadata.refreshToken,
-      };
-
-      // Stub the http request (OAuth2.requestToken())
-      _postParmsStub.returns(Promise.resolve(authResponse));
-      const userInfoResponseBody = {
-        statusCode: 404,
-        body: JSON.stringify([
+        stubUserRequest(
+          { body: { preferred_username: testOrg.username, organization_id: testOrg.orgId }, statusCode: 200 },
           {
-            message: 'Could not retrieve the username after successful auth code exchange.\nDue to: %s',
-            errorCode: 'AuthCodeUsernameRetrievalError',
-          },
-        ]),
-      };
-      const userResponseBody = {
-        body: JSON.stringify({ Username: testMetadata.username.toUpperCase() }),
-      };
-      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest')
-        .onFirstCall()
-        .returns(Promise.resolve(userInfoResponseBody))
-        .onSecondCall()
-        .returns(Promise.resolve(userResponseBody));
+            statusCode: 404,
+            body: [
+              {
+                message: 'Could not retrieve the username after successful auth code exchange.\nDue to: %s',
+                errorCode: 'AuthCodeUsernameRetrievalError',
+              },
+            ],
+          }
+        );
 
-      // Create the auth code AuthInfo instance
-      try {
-        await AuthInfo.create({ oauth2Options: authCodeConfig });
-        assert.fail('should have thrown an error within AuthInfo.buildWebAuthConfig()');
-      } catch (err) {
-        expect(err.name).to.equal('AuthCodeUsernameRetrievalError');
-      }
-    });
-    it('should throw a AuthCodeUsernameRetrievalError when userInfo retrieval fails after auth code exchange', async () => {
-      const authCodeConfig = {
-        authCode: testMetadata.authCode,
-        loginUrl: testMetadata.loginUrl,
-      };
-      const authResponse = {
-        access_token: testMetadata.accessToken,
-        instance_url: testMetadata.instanceUrl,
-        id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
-        refresh_token: testMetadata.refreshToken,
-      };
+        // Create the auth code AuthInfo instance
+        try {
+          await AuthInfo.create({ oauth2Options: authCodeConfig });
+          assert.fail('should have thrown an error within AuthInfo.buildWebAuthConfig()');
+        } catch (err) {
+          expect(err.name).to.equal('AuthCodeUsernameRetrievalError');
+        }
+      });
 
-      // Stub the http request (OAuth2.requestToken())
-      _postParmsStub.returns(Promise.resolve(authResponse));
-      const userInfoResponseBody = {
-        statusCode: 404,
-        body: JSON.stringify([
-          {
-            message: 'Could not retrieve the username after successful auth code exchange.\nDue to: %s',
-            errorCode: 'AuthCodeUsernameRetrievalError',
-          },
-        ]),
-      };
-      const userResponseBody = {
-        body: JSON.stringify({ Username: testMetadata.username.toUpperCase() }),
-      };
-      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest')
-        .onFirstCall()
-        .returns(Promise.resolve(userInfoResponseBody))
-        .onSecondCall()
-        .returns(Promise.resolve(userResponseBody));
-
-      // Create the auth code AuthInfo instance
-      try {
-        await AuthInfo.create({ oauth2Options: authCodeConfig });
-        assert.fail('should have thrown an error within AuthInfo.buildWebAuthConfig()');
-      } catch (err) {
-        expect(err.name).to.equal('AuthCodeUsernameRetrievalError');
-      }
-    });
-
-    it('should throw a AuthCodeUsernameRetrievalError when user sobject retrieval fails after auth code exchange', async () => {
-      const authCodeConfig = {
-        authCode: testMetadata.authCode,
-        loginUrl: testMetadata.loginUrl,
-      };
-      const authResponse = {
-        access_token: testMetadata.accessToken,
-        instance_url: testMetadata.instanceUrl,
-        id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
-        refresh_token: testMetadata.refreshToken,
-      };
-
-      // Stub the http request (OAuth2.requestToken())
-      _postParmsStub.returns(Promise.resolve(authResponse));
-      const userInfoResponseBody = {
-        body: JSON.stringify({ preferred_username: testMetadata.username, organization_id: testMetadata.orgId }),
-      };
-      const userResponseBody = {
-        statusCode: 404,
-        body: JSON.stringify([
-          {
-            message: 'Could not retrieve the username after successful auth code exchange.\nDue to: %s',
-            errorCode: 'AuthCodeUsernameRetrievalError',
-          },
-        ]),
-      };
-      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest')
-        .onFirstCall()
-        .returns(Promise.resolve(userInfoResponseBody))
-        .onSecondCall()
-        .returns(Promise.resolve(userResponseBody));
-
-      // Create the auth code AuthInfo instance
-      try {
-        await AuthInfo.create({ oauth2Options: authCodeConfig });
-        assert.fail('should have thrown an error within AuthInfo.buildWebAuthConfig()');
-      } catch (err) {
-        expect(err.name).to.equal('AuthCodeUsernameRetrievalError');
-      }
-    });
-
-    it('should throw an error when neither username nor options have been passed', async () => {
-      try {
-        await AuthInfo.create();
-        assert.fail('Expected AuthInfo.create() to throw an error when no params are passed');
-      } catch (err) {
-        expect(err.name).to.equal('AuthInfoCreationError');
-      }
+      it('should throw an error when neither username nor options have been passed', async () => {
+        try {
+          await AuthInfo.create();
+          assert.fail('Expected AuthInfo.create() to throw an error when no params are passed');
+        } catch (err) {
+          expect(err.name).to.equal('AuthInfoCreationError');
+        }
+      });
     });
   });
 
-  describe('save()', () => {
+  describe('save', () => {
     it('should update the AuthInfo fields, and write to file', async () => {
-      const username = 'authInfoTest_username_SaveTest1';
       const refreshTokenConfig = {
-        refreshToken: testMetadata.refreshToken,
-        loginUrl: testMetadata.loginUrl,
+        refreshToken: testOrg.refreshToken,
+        loginUrl: testOrg.loginUrl,
       };
       const authResponse = {
-        access_token: testMetadata.accessToken,
-        instance_url: testMetadata.instanceUrl,
+        access_token: testOrg.accessToken,
+        instance_url: testOrg.instanceUrl,
         id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
-        expirationDate: testMetadata.expirationDate,
+        expirationDate: testOrg.expirationDate,
       };
 
       // Stub the http request (OAuth2.refreshToken())
-      _postParmsStub.returns(Promise.resolve(authResponse));
+      postParamsStub.resolves(authResponse);
 
       // Create the AuthInfo instance
       const authInfo = await AuthInfo.create({
-        username,
+        username: testOrg.username,
         oauth2Options: refreshTokenConfig,
       });
 
-      expect(authInfo.getUsername()).to.equal(username);
+      expect(authInfo.getUsername()).to.equal(testOrg.username);
 
       // reset the AuthInfo.update stub so we only look at what happens with AuthInfo.save().
-      authInfoUpdate.resetHistory();
+      authInfoStubs.update.resetHistory();
 
       // Save new fields
-      const changedData = { accessToken: testMetadata.accessToken, expirationDate: testMetadata.expirationDate };
-
-      stubMethod($$.SANDBOX, testMetadata, 'fetchConfigInfo').returns(Promise.resolve({}));
+      const changedData = { accessToken: testOrg.accessToken, expirationDate: testOrg.expirationDate };
       await authInfo.save(changedData);
 
-      expect(authInfoUpdate.called).to.be.true;
-      expect(authInfoUpdate.firstCall.args[0]).to.deep.equal(changedData);
-      expect(configFileWrite.called).to.be.true;
+      expect(authInfoStubs.update.called).to.be.true;
+      expect(authInfoStubs.update.firstCall.args[0]).to.deep.equal(changedData);
+      // expect(configFileWrite.called).to.be.true;
 
       const crypto = await Crypto.create();
-      const decryptedActualFields = configFileWrite.lastCall.thisValue.toObject();
+      // const decryptedActualFields = configFileWrite.lastCall.thisValue.toObject();
+      const decryptedActualFields = $$.stubs.configWrite.lastCall.thisValue.toObject();
       decryptedActualFields.accessToken = crypto.decrypt(decryptedActualFields.accessToken);
       decryptedActualFields.refreshToken = crypto.decrypt(decryptedActualFields.refreshToken);
       decryptedActualFields.clientSecret = crypto.decrypt(decryptedActualFields.clientSecret);
       delete decryptedActualFields.timestamp;
       const expectedFields = {
         accessToken: changedData.accessToken,
-        instanceUrl: testMetadata.instanceUrl,
-        username,
+        instanceUrl: testOrg.instanceUrl,
+        username: testOrg.username,
         orgId: authResponse.id.split('/')[0],
         loginUrl: refreshTokenConfig.loginUrl,
         refreshToken: refreshTokenConfig.refreshToken,
@@ -1245,7 +1033,7 @@ describe('AuthInfo', () => {
         // We just hard code the legacy values here to ensure old auth files will still work.
         clientId: 'SalesforceDevelopmentExperience',
         clientSecret: '1384510088588713504',
-        expirationDate: testMetadata.expirationDate,
+        expirationDate: testOrg.expirationDate,
       };
       // Note that this also verifies the clientId and clientSecret are not persisted,
       // and that data is encrypted when saved (because we have to decrypt it to verify here).
@@ -1262,37 +1050,37 @@ describe('AuthInfo', () => {
         username,
         accessTokenOptions: {
           accessToken: username,
-          instanceUrl: testMetadata.instanceUrl,
+          instanceUrl: testOrg.instanceUrl,
         },
       });
 
       expect(authInfo.getUsername()).to.equal(username);
 
-      configFileWrite.rejects(new Error('Should not call save'));
+      $$.stubs.configWrite.rejects(new Error('Should not call save'));
       await authInfo.save();
       // If the test doesn't blow up, it is a success because the write (reject) never happened
     });
   });
 
-  describe('refreshFn()', () => {
+  describe('refreshFn', () => {
     it('should call init() and save()', async () => {
       const context = {
         getUsername: () => '',
         getFields: (decrypt = false) => ({
-          loginUrl: testMetadata.loginUrl,
-          clientId: testMetadata.clientId,
+          loginUrl: testOrg.loginUrl,
+          clientId: testOrg.clientId,
           privateKey: 'authInfoTest/jwt/server.key',
-          accessToken: decrypt ? testMetadata.accessToken : testMetadata.encryptedAccessToken,
+          accessToken: decrypt ? testOrg.accessToken : testOrg.encryptedAccessToken,
         }),
         initAuthOptions: $$.SANDBOX.stub(),
         save: $$.SANDBOX.stub(),
         logger: $$.TEST_LOGGER,
       };
       const testCallback = $$.SANDBOX.stub();
-      testCallback.returns(Promise.resolve());
+      testCallback.resolves();
 
-      context.initAuthOptions.returns(Promise.resolve());
-      context.save.returns(Promise.resolve());
+      context.initAuthOptions.resolves();
+      context.save.resolves();
       // @ts-ignore
       await AuthInfo.prototype['refreshFn'].call(context, null, testCallback);
 
@@ -1302,18 +1090,18 @@ describe('AuthInfo', () => {
         loginUrl: context.getFields().loginUrl,
         clientId: context.getFields().clientId,
         privateKey: context.getFields().privateKey,
-        accessToken: testMetadata.accessToken,
+        accessToken: testOrg.accessToken,
       };
       expect(context.initAuthOptions.firstCall.args[0]).to.deep.equal(expectedInitArgs);
       expect(context.save.called, 'Should have called AuthInfo.save() during refreshFn()').to.be.true;
       expect(testCallback.called, 'Should have called the callback passed to refreshFn()').to.be.true;
-      expect(testCallback.firstCall.args[1]).to.equal(testMetadata.accessToken);
+      expect(testCallback.firstCall.args[1]).to.equal(testOrg.accessToken);
     });
 
     it('should path.resolve jwtkeyfilepath', async () => {
-      const pathSpy = $$.SANDBOX.spy(pathImport, 'resolve');
+      const resolveSpy = $$.SANDBOX.spy(pathImport, 'resolve');
 
-      authInfoBuildJwtConfig.restore();
+      authInfoStubs.authJwt.restore();
       stubMethod($$.SANDBOX, AuthInfo.prototype, 'authJwt').resolves({
         instanceUrl: '',
         accessToken: '',
@@ -1321,23 +1109,23 @@ describe('AuthInfo', () => {
       stubMethod($$.SANDBOX, AuthInfo.prototype, 'determineIfDevHub').resolves(false);
 
       await AuthInfo.create({
-        username: 'cristiand391',
+        username: testOrg.username,
         oauth2Options: {
-          clientId: '1234',
-          privateKeyFile: 'authInfoTest/jwt/server.key',
+          clientId: testOrg.clientId,
+          privateKeyFile: testOrg.privateKey,
         },
       });
-      expect(pathSpy.lastCall.args[0]).to.equal('authInfoTest/jwt/server.key');
+      expect(resolveSpy.lastCall.args[0]).to.equal(testOrg.privateKey);
     });
 
     it('should call the callback with OrgDataNotAvailableError when AuthInfo.init() fails', async () => {
       const context = {
         getUsername: () => '',
         getFields: () => ({
-          loginUrl: testMetadata.loginUrl,
-          clientId: testMetadata.clientId,
-          privateKey: 'authInfoTest/jwt/server.key',
-          accessToken: testMetadata.encryptedAccessToken,
+          loginUrl: testOrg.loginUrl,
+          clientId: testOrg.clientId,
+          privateKey: testOrg.privateKey,
+          accessToken: testOrg.encryptedAccessToken,
         }),
         initAuthOptions: $$.SANDBOX.stub(),
         save: $$.SANDBOX.stub(),
@@ -1345,8 +1133,7 @@ describe('AuthInfo', () => {
       };
       const testCallback = $$.SANDBOX.spy();
       context.initAuthOptions.throws(new Error('Error: Data Not Available'));
-      context.save.returns(Promise.resolve());
-      // @ts-ignore
+      context.save.resolves();
       await AuthInfo.prototype['refreshFn'].call(context, null, testCallback);
       expect(testCallback.called).to.be.true;
       const sfError = testCallback.firstCall.args[0];
@@ -1354,7 +1141,7 @@ describe('AuthInfo', () => {
     });
   });
 
-  describe('getAuthorizationUrl()', () => {
+  describe('getAuthorizationUrl', () => {
     let scope: string;
     beforeEach(() => {
       scope = env.getString('SFDX_AUTH_SCOPES', '');
@@ -1365,11 +1152,11 @@ describe('AuthInfo', () => {
 
     it('should return the correct url', () => {
       const options = {
-        clientId: testMetadata.clientId,
-        redirectUri: testMetadata.redirectUri,
-        loginUrl: testMetadata.loginUrl,
+        clientId: testOrg.clientId,
+        redirectUri: testOrg.redirectUri,
+        loginUrl: testOrg.loginUrl,
       };
-      const url: string = AuthInfo.getAuthorizationUrl.call(null, options);
+      const url = AuthInfo.getAuthorizationUrl(options);
 
       expect(url.startsWith(options.loginUrl), 'authorization URL should start with the loginUrl').to.be.true;
       expect(url).to.contain('state=');
@@ -1379,12 +1166,12 @@ describe('AuthInfo', () => {
 
     it('should return the correct url with modified scope', () => {
       const options = {
-        clientId: testMetadata.clientId,
-        redirectUri: testMetadata.redirectUri,
-        loginUrl: testMetadata.loginUrl,
+        clientId: testOrg.clientId,
+        redirectUri: testOrg.redirectUri,
+        loginUrl: testOrg.loginUrl,
         scope: 'test',
       };
-      const url: string = AuthInfo.getAuthorizationUrl.call(null, options);
+      const url = AuthInfo.getAuthorizationUrl(options);
 
       expect(url.startsWith(options.loginUrl), 'authorization URL should start with the loginUrl').to.be.true;
       expect(url).to.contain('state=');
@@ -1395,11 +1182,11 @@ describe('AuthInfo', () => {
     it('should return the correct url with env scope', () => {
       env.setString('SFDX_AUTH_SCOPES', 'from-env');
       const options = {
-        clientId: testMetadata.clientId,
-        redirectUri: testMetadata.redirectUri,
-        loginUrl: testMetadata.loginUrl,
+        clientId: testOrg.clientId,
+        redirectUri: testOrg.redirectUri,
+        loginUrl: testOrg.loginUrl,
       };
-      const url: string = AuthInfo.getAuthorizationUrl.call(null, options);
+      const url = AuthInfo.getAuthorizationUrl(options);
 
       expect(url.startsWith(options.loginUrl), 'authorization URL should start with the loginUrl').to.be.true;
       expect(url).to.contain('state=');
@@ -1410,12 +1197,12 @@ describe('AuthInfo', () => {
     it('should return the correct url with option over env', () => {
       env.setString('SFDX_AUTH_SCOPES', 'from-env');
       const options = {
-        clientId: testMetadata.clientId,
-        redirectUri: testMetadata.redirectUri,
-        loginUrl: testMetadata.loginUrl,
+        clientId: testOrg.clientId,
+        redirectUri: testOrg.redirectUri,
+        loginUrl: testOrg.loginUrl,
         scope: 'from-option',
       };
-      const url: string = AuthInfo.getAuthorizationUrl.call(null, options);
+      const url = AuthInfo.getAuthorizationUrl(options);
 
       expect(url.startsWith(options.loginUrl), 'authorization URL should start with the loginUrl').to.be.true;
       expect(url).to.contain('state=');
@@ -1424,83 +1211,77 @@ describe('AuthInfo', () => {
     });
   });
 
-  describe('getSfdxAuthUrl()', () => {
+  describe('getSfdxAuthUrl', () => {
     it('should return the correct sfdx auth url', async () => {
-      const username = 'authInfoTest_username_RefreshToken';
-      const refreshTokenConfig = {
-        refreshToken: testMetadata.refreshToken,
-        loginUrl: testMetadata.loginUrl,
-      };
       const authResponse = {
-        access_token: testMetadata.accessToken,
-        instance_url: testMetadata.instanceUrl,
+        access_token: testOrg.accessToken,
+        instance_url: testOrg.instanceUrl,
         id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
       };
 
       // Stub the http request (OAuth2.refreshToken())
-      _postParmsStub.returns(Promise.resolve(authResponse));
+      postParamsStub.resolves(authResponse);
 
       // Create the refresh token AuthInfo instance
       const authInfo = await AuthInfo.create({
-        username,
-        oauth2Options: refreshTokenConfig,
+        username: testOrg.username,
+        oauth2Options: {
+          refreshToken: testOrg.refreshToken,
+          loginUrl: testOrg.loginUrl,
+        },
       });
 
       expect(authInfo.getSfdxAuthUrl()).to.contain(
-        `force://SalesforceDevelopmentExperience:1384510088588713504:${testMetadata.refreshToken}@mydevhub.localhost.internal.salesforce.com:6109`
+        `force://SalesforceDevelopmentExperience:1384510088588713504:${
+          testOrg.refreshToken
+        }@${testOrg.instanceUrl.replace('https://', '')}`
       );
     });
 
     it('should handle undefined client secret', async () => {
-      const username = 'authInfoTest_username_RefreshToken';
-      const refreshTokenConfig = {
-        refreshToken: testMetadata.refreshToken,
-        loginUrl: testMetadata.loginUrl,
-      };
-
       const authResponse = {
-        access_token: testMetadata.accessToken,
-        instance_url: testMetadata.instanceUrl,
+        access_token: testOrg.accessToken,
+        instance_url: testOrg.instanceUrl,
         id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
       };
 
       // Stub the http request (OAuth2.refreshToken())
-      _postParmsStub.returns(Promise.resolve(authResponse));
+      postParamsStub.resolves(authResponse);
 
       // Create the refresh token AuthInfo instance
       const authInfo = await AuthInfo.create({
-        username,
-        oauth2Options: refreshTokenConfig,
+        username: testOrg.username,
+        oauth2Options: {
+          refreshToken: testOrg.refreshToken,
+          loginUrl: testOrg.loginUrl,
+        },
       });
 
       // delete the client secret
       delete authInfo.getFields().clientSecret;
-
+      const instanceUrl = testOrg.instanceUrl.replace('https://', '');
       expect(authInfo.getSfdxAuthUrl()).to.contain(
-        `force://SalesforceDevelopmentExperience::${testMetadata.refreshToken}@mydevhub.localhost.internal.salesforce.com:6109`
+        `force://SalesforceDevelopmentExperience::${testOrg.refreshToken}@${instanceUrl}`
       );
     });
 
     it('should handle undefined refresh token', async () => {
-      const username = 'authInfoTest_username_RefreshToken';
-      const refreshTokenConfig = {
-        refreshToken: testMetadata.refreshToken,
-        loginUrl: testMetadata.loginUrl,
-      };
-
       const authResponse = {
-        access_token: testMetadata.accessToken,
-        instance_url: testMetadata.instanceUrl,
+        access_token: testOrg.accessToken,
+        instance_url: testOrg.instanceUrl,
         id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
       };
 
       // Stub the http request (OAuth2.refreshToken())
-      _postParmsStub.returns(Promise.resolve(authResponse));
+      postParamsStub.resolves(authResponse);
 
       // Create the refresh token AuthInfo instance
       const authInfo = await AuthInfo.create({
-        username,
-        oauth2Options: refreshTokenConfig,
+        username: testOrg.username,
+        oauth2Options: {
+          refreshToken: testOrg.refreshToken,
+          loginUrl: testOrg.loginUrl,
+        },
       });
 
       // delete the refresh token
@@ -1510,25 +1291,22 @@ describe('AuthInfo', () => {
     });
 
     it('should handle undefined instance url', async () => {
-      const username = 'authInfoTest_username_RefreshToken';
-      const refreshTokenConfig = {
-        refreshToken: testMetadata.refreshToken,
-        loginUrl: testMetadata.loginUrl,
-      };
-
       const authResponse = {
-        access_token: testMetadata.accessToken,
-        instance_url: testMetadata.instanceUrl,
+        access_token: testOrg.accessToken,
+        instance_url: testOrg.instanceUrl,
         id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
       };
 
       // Stub the http request (OAuth2.refreshToken())
-      _postParmsStub.returns(Promise.resolve(authResponse));
+      postParamsStub.resolves(authResponse);
 
       // Create the refresh token AuthInfo instance
       const authInfo = await AuthInfo.create({
-        username,
-        oauth2Options: refreshTokenConfig,
+        username: testOrg.username,
+        oauth2Options: {
+          refreshToken: testOrg.refreshToken,
+          loginUrl: testOrg.loginUrl,
+        },
       });
 
       // delete the instance url
@@ -1539,21 +1317,18 @@ describe('AuthInfo', () => {
   });
 
   describe('setAlias', () => {
-    const username = 'authInfoTest_username';
     const alias = 'MyAlias';
 
     it('should set alias', async () => {
       const aliasAccessorSpy = spyMethod($$.SANDBOX, AliasAccessor.prototype, 'set');
-      const authInfo = await AuthInfo.create({ username });
+      const authInfo = await AuthInfo.create({ username: testOrg.username });
       await authInfo.setAlias(alias);
       expect(aliasAccessorSpy.calledOnce).to.be.true;
-      expect(aliasAccessorSpy.firstCall.args[0]).to.equal(alias);
-      expect(aliasAccessorSpy.firstCall.args[1]).to.equal(username);
+      expect(aliasAccessorSpy.firstCall.args).to.deep.equal([alias, testOrg.username]);
     });
   });
 
   describe('setAsDefault', () => {
-    const username = 'authInfoTest_username';
     const alias = 'MyAlias';
     let configSpy: sinon.SinonSpy;
 
@@ -1563,23 +1338,23 @@ describe('AuthInfo', () => {
 
     it('should set username to target-org', async () => {
       stubMethod($$.SANDBOX, AliasAccessor.prototype, 'get').returns(null);
-      const authInfo = await AuthInfo.create({ username });
+      const authInfo = await AuthInfo.create({ username: testOrg.username });
       await authInfo.setAsDefault({ org: true });
       expect(configSpy.called).to.be.true;
-      expect(configSpy.firstCall.args).to.deep.equal([OrgConfigProperties.TARGET_ORG, username]);
+      expect(configSpy.firstCall.args).to.deep.equal([OrgConfigProperties.TARGET_ORG, testOrg.username]);
     });
 
     it('should set username to target-dev-hub', async () => {
       stubMethod($$.SANDBOX, AliasAccessor.prototype, 'get').returns(null);
-      const authInfo = await AuthInfo.create({ username });
+      const authInfo = await AuthInfo.create({ username: testOrg.username });
       await authInfo.setAsDefault({ devHub: true });
       expect(configSpy.called).to.be.true;
-      expect(configSpy.firstCall.args).to.deep.equal([OrgConfigProperties.TARGET_DEV_HUB, username]);
+      expect(configSpy.firstCall.args).to.deep.equal([OrgConfigProperties.TARGET_DEV_HUB, testOrg.username]);
     });
 
     it('should set alias to target-org', async () => {
       stubMethod($$.SANDBOX, AliasAccessor.prototype, 'get').returns(alias);
-      const authInfo = await AuthInfo.create({ username });
+      const authInfo = await AuthInfo.create({ username: testOrg.username });
       await authInfo.setAsDefault({ org: true });
       expect(configSpy.called).to.be.true;
       expect(configSpy.firstCall.args).to.deep.equal([OrgConfigProperties.TARGET_ORG, alias]);
@@ -1587,85 +1362,58 @@ describe('AuthInfo', () => {
 
     it('should set alias to target-dev-hub', async () => {
       stubMethod($$.SANDBOX, AliasAccessor.prototype, 'get').returns(alias);
-      const authInfo = await AuthInfo.create({ username });
+      const authInfo = await AuthInfo.create({ username: testOrg.username });
       await authInfo.setAsDefault({ devHub: true });
       expect(configSpy.called).to.be.true;
       expect(configSpy.firstCall.args).to.deep.equal([OrgConfigProperties.TARGET_DEV_HUB, alias]);
+    });
+
+    it('should use global config if local config fails', async () => {
+      stubMethod($$.SANDBOX, AliasAccessor.prototype, 'get').returns(null);
+      stubMethod($$.SANDBOX, Config, 'create')
+        .withArgs({ isGlobal: false })
+        .throws()
+        .withArgs({ isGlobal: true })
+        .callThrough();
+      const authInfo = await AuthInfo.create({ username: testOrg.username });
+      await authInfo.setAsDefault({ org: true });
+      expect(configSpy.called).to.be.true;
+      expect(configSpy.firstCall.args).to.deep.equal([OrgConfigProperties.TARGET_ORG, testOrg.username]);
     });
   });
 
   describe('getDefaultInstanceUrl', () => {
     it('should return the configured instance url if it exists', async () => {
-      stubMethod($$.SANDBOX, ConfigAggregator, 'getValue').returns({ value: testMetadata.instanceUrl });
-      const result = AuthInfo.getDefaultInstanceUrl();
-      expect(result).to.equal(testMetadata.instanceUrl);
+      $$.stubConfig({ [OrgConfigProperties.ORG_INSTANCE_URL]: testOrg.instanceUrl });
+      expect(AuthInfo.getDefaultInstanceUrl()).to.equal(testOrg.instanceUrl);
     });
 
     it('should return the default instance url if no configured instance url exists', async () => {
-      stubMethod($$.SANDBOX, ConfigAggregator, 'getValue').returns({ value: null });
-      const result = AuthInfo.getDefaultInstanceUrl();
-      expect(result).to.equal('https://login.salesforce.com');
+      expect(AuthInfo.getDefaultInstanceUrl()).to.equal('https://login.salesforce.com');
     });
   });
 
   describe('hasAuthentications', () => {
     it('should return false', async () => {
       stubMethod($$.SANDBOX, OrgAccessor.prototype, 'list').returns([]);
-      const result = await AuthInfo.hasAuthentications();
-      expect(result).to.be.false;
+      expect(await AuthInfo.hasAuthentications()).to.be.false;
     });
 
     it('should return true', async () => {
-      await $$.stubAuths(new MockTestOrgData());
-      const result = await AuthInfo.hasAuthentications();
-      expect(result).to.be.equal(true);
+      await $$.stubAuths(testOrg);
+      expect(await AuthInfo.hasAuthentications()).to.be.true;
     });
   });
 
   describe('listAllAuthorizations', () => {
     describe('with no AuthInfo.create errors', () => {
-      const username = 'espresso@coffee.com';
-      let authInfo: AuthInfo;
       beforeEach(async () => {
-        stubMethod($$.SANDBOX, ConfigAggregator.prototype, 'loadProperties').callsFake(async () => {});
-        stubMethod($$.SANDBOX, ConfigAggregator.prototype, 'getPropertyValue').returns(testMetadata.instanceUrl);
-        // Stub the http request (OAuth2.refreshToken())
-        // This will be called for both, and we want to make sure the clientSecrete is the
-        // same for both.
-        _postParmsStub.callsFake((params) => {
-          expect(params.client_secret).to.deep.equal(testMetadata.clientSecret);
-          return {
-            access_token: testMetadata.accessToken,
-            instance_url: testMetadata.instanceUrl,
-            refresh_token: testMetadata.refreshToken,
-            id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
-          };
-        });
-
-        const jwtData = {};
-        set(jwtData, 'accessToken', testMetadata.encryptedAccessToken);
-        set(jwtData, 'clientId', testMetadata.clientId);
-        set(jwtData, 'loginUrl', testMetadata.loginUrl);
-        set(jwtData, 'instanceUrl', testMetadata.instanceUrl);
-        set(jwtData, 'privateKey', 'authInfoTest/jwt/server.key');
-        set(jwtData, 'username', username);
-        testMetadata.fetchConfigInfo = () => jwtData;
-
-        authInfo = await AuthInfo.create({
-          username,
-          oauth2Options: {
-            clientId: testMetadata.clientId,
-            clientSecret: testMetadata.clientSecret,
-            loginUrl: testMetadata.instanceUrl,
-            authCode: testMetadata.authCode,
-          },
-        });
-        stubMethod($$.SANDBOX, AuthInfo, 'create').withArgs({ username }).returns(Promise.resolve(authInfo));
-        stubMethod($$.SANDBOX, OrgAccessor.prototype, 'readAll').resolves([authInfo.getFields()]);
+        await $$.stubAuths(testOrg);
       });
 
       it('should return list of authorizations with web oauthMethod', async () => {
-        stubMethod($$.SANDBOX, AliasAccessor.prototype, 'getAll').returns([]);
+        stubMethod($$.SANDBOX, AuthInfo.prototype, 'isJwt').returns(false);
+        stubMethod($$.SANDBOX, AuthInfo.prototype, 'isOauth').returns(true);
         const auths = await AuthInfo.listAllAuthorizations(
           (orgAuth) => orgAuth.oauthMethod !== 'jwt' && !orgAuth.isScratchOrg
         );
@@ -1674,10 +1422,10 @@ describe('AuthInfo', () => {
             aliases: [],
             configs: [],
             isScratchOrg: false,
-            username: 'espresso@coffee.com',
-            orgId: '00DAuthInfoTest_orgId',
-            instanceUrl: 'https://mydevhub.localhost.internal.salesforce.com:6109',
-            accessToken: 'authInfoTest_access_token',
+            username: testOrg.username,
+            orgId: testOrg.orgId,
+            instanceUrl: testOrg.instanceUrl,
+            accessToken: testOrg.accessToken,
             oauthMethod: 'web',
             isDevHub: false,
             isExpired: 'unknown',
@@ -1687,21 +1435,17 @@ describe('AuthInfo', () => {
       });
 
       it('should return list of authorizations with jwt oauthMethod', async () => {
-        stubMethod($$.SANDBOX, AliasAccessor.prototype, 'getAll').returns([]);
         stubMethod($$.SANDBOX, AuthInfo.prototype, 'isJwt').returns(true);
         const auths = await AuthInfo.listAllAuthorizations();
-        const expiryDate = new Date(Date.now());
-        expiryDate.setFullYear(expiryDate.getFullYear() - 1);
-        authInfo.getFields().expirationDate = expiryDate.toISOString();
         expect(auths).to.deep.equal([
           {
             aliases: [],
             configs: [],
             isScratchOrg: false,
-            username: 'espresso@coffee.com',
-            orgId: '00DAuthInfoTest_orgId',
-            instanceUrl: 'https://mydevhub.localhost.internal.salesforce.com:6109',
-            accessToken: 'authInfoTest_access_token',
+            username: testOrg.username,
+            orgId: testOrg.orgId,
+            instanceUrl: testOrg.instanceUrl,
+            accessToken: testOrg.accessToken,
             oauthMethod: 'jwt',
             isDevHub: false,
             isExpired: 'unknown',
@@ -1711,7 +1455,6 @@ describe('AuthInfo', () => {
       });
 
       it('should return list of authorizations with token oauthMethod', async () => {
-        stubMethod($$.SANDBOX, AliasAccessor.prototype, 'getAll').returns([]);
         stubMethod($$.SANDBOX, AuthInfo.prototype, 'isJwt').returns(false);
         stubMethod($$.SANDBOX, AuthInfo.prototype, 'isOauth').returns(false);
         const auths = await AuthInfo.listAllAuthorizations();
@@ -1720,10 +1463,10 @@ describe('AuthInfo', () => {
             aliases: [],
             configs: [],
             isScratchOrg: false,
-            username: 'espresso@coffee.com',
-            orgId: '00DAuthInfoTest_orgId',
-            instanceUrl: 'https://mydevhub.localhost.internal.salesforce.com:6109',
-            accessToken: 'authInfoTest_access_token',
+            username: testOrg.username,
+            orgId: testOrg.orgId,
+            instanceUrl: testOrg.instanceUrl,
+            accessToken: testOrg.accessToken,
             oauthMethod: 'token',
             isDevHub: false,
             isExpired: 'unknown',
@@ -1733,7 +1476,7 @@ describe('AuthInfo', () => {
       });
 
       it('should return list of authorizations with aliases', async () => {
-        stubMethod($$.SANDBOX, AliasAccessor.prototype, 'getAll').returns(['MyAlias']);
+        $$.stubAliases({ MyAlias: testOrg.username });
         const auths = await AuthInfo.listAllAuthorizations(
           (orgAuth) => orgAuth.aliases.length === 1 && orgAuth.aliases.includes('MyAlias')
         );
@@ -1742,11 +1485,11 @@ describe('AuthInfo', () => {
             aliases: ['MyAlias'],
             configs: [],
             isScratchOrg: false,
-            username: 'espresso@coffee.com',
-            orgId: '00DAuthInfoTest_orgId',
-            instanceUrl: 'https://mydevhub.localhost.internal.salesforce.com:6109',
-            accessToken: 'authInfoTest_access_token',
-            oauthMethod: 'web',
+            username: testOrg.username,
+            orgId: testOrg.orgId,
+            instanceUrl: testOrg.instanceUrl,
+            accessToken: testOrg.accessToken,
+            oauthMethod: 'jwt',
             isDevHub: false,
             isExpired: 'unknown',
             isSandbox: false,
@@ -1755,30 +1498,24 @@ describe('AuthInfo', () => {
       });
 
       it('should return list of authorizations with configs', async () => {
-        stubMethod($$.SANDBOX, AliasAccessor.prototype, 'getAll').returns(['MyAlias']);
-        stubMethod($$.SANDBOX, ConfigAggregator.prototype, 'getConfigInfo').returns([
-          {
-            value: 'MyAlias',
-            key: OrgConfigProperties.TARGET_ORG,
-          },
-          {
-            value: username,
-            key: OrgConfigProperties.TARGET_DEV_HUB,
-          },
-        ]);
+        $$.stubAliases({ MyAlias: testOrg.username });
+        $$.stubConfig({
+          [OrgConfigProperties.TARGET_ORG]: 'MyAlias',
+          [OrgConfigProperties.TARGET_DEV_HUB]: testOrg.username,
+        });
         const auths = await AuthInfo.listAllAuthorizations((orgAuth) =>
           orgAuth.configs.includes(OrgConfigProperties.TARGET_ORG)
         );
         expect(auths).to.deep.equal([
           {
             aliases: ['MyAlias'],
-            configs: [OrgConfigProperties.TARGET_ORG, OrgConfigProperties.TARGET_DEV_HUB],
+            configs: [OrgConfigProperties.TARGET_DEV_HUB, OrgConfigProperties.TARGET_ORG],
             isScratchOrg: false,
-            username: 'espresso@coffee.com',
-            orgId: '00DAuthInfoTest_orgId',
-            instanceUrl: 'https://mydevhub.localhost.internal.salesforce.com:6109',
-            accessToken: 'authInfoTest_access_token',
-            oauthMethod: 'web',
+            username: testOrg.username,
+            orgId: testOrg.orgId,
+            instanceUrl: testOrg.instanceUrl,
+            accessToken: testOrg.accessToken,
+            oauthMethod: 'jwt',
             isDevHub: false,
             isExpired: 'unknown',
             isSandbox: false,
@@ -1787,22 +1524,24 @@ describe('AuthInfo', () => {
       });
 
       it('should return list of authorizations devhub username', async () => {
-        stubMethod($$.SANDBOX, AliasAccessor.prototype, 'getAll').returns([]);
-        authInfo.getFields().devHubUsername = 'foobarusername';
         const expiryDate = new Date(Date.now());
         expiryDate.setFullYear(expiryDate.getFullYear() + 1);
-        authInfo.getFields().expirationDate = expiryDate.toISOString();
+        stubMethod($$.SANDBOX, AuthInfo.prototype, 'getFields').returns({
+          ...(await testOrg.getConfig()),
+          devHubUsername: 'foobarusername',
+          expirationDate: expiryDate.toISOString(),
+        });
         const auths = await AuthInfo.listAllAuthorizations();
         expect(auths).to.deep.equal([
           {
             aliases: [],
             configs: [],
             isScratchOrg: true,
-            username: 'espresso@coffee.com',
-            orgId: '00DAuthInfoTest_orgId',
-            instanceUrl: 'https://mydevhub.localhost.internal.salesforce.com:6109',
-            accessToken: 'authInfoTest_access_token',
-            oauthMethod: 'web',
+            username: testOrg.username,
+            orgId: testOrg.orgId,
+            instanceUrl: testOrg.instanceUrl,
+            accessToken: testOrg.accessToken,
+            oauthMethod: 'jwt',
             isDevHub: false,
             isExpired: false,
             isSandbox: false,
@@ -1813,33 +1552,19 @@ describe('AuthInfo', () => {
 
     describe('with AuthInfo.create errors', () => {
       beforeEach(async () => {
-        const username = 'espresso@coffee.com';
-        stubMethod($$.SANDBOX, ConfigAggregator.prototype, 'loadProperties').callsFake(async () => {});
-        stubMethod($$.SANDBOX, ConfigAggregator.prototype, 'getPropertyValue').returns(testMetadata.instanceUrl);
-
-        const jwtData = {};
-        set(jwtData, 'accessToken', testMetadata.encryptedAccessToken);
-        set(jwtData, 'clientId', testMetadata.clientId);
-        set(jwtData, 'loginUrl', testMetadata.loginUrl);
-        set(jwtData, 'instanceUrl', testMetadata.instanceUrl);
-        set(jwtData, 'privateKey', 'authInfoTest/jwt/server.key');
-        set(jwtData, 'orgId', '00DAuthInfoTest_orgId');
-        set(jwtData, 'username', username);
-        testMetadata.fetchConfigInfo = () => jwtData;
-        stubMethod($$.SANDBOX, AuthInfo, 'create').withArgs({ username }).throws(new Error('FAIL!'));
-        stubMethod($$.SANDBOX, OrgAccessor.prototype, 'readAll').resolves([jwtData]);
+        await $$.stubAuths(testOrg);
+        stubMethod($$.SANDBOX, AuthInfo, 'create').withArgs({ username: testOrg.username }).throws(new Error('FAIL!'));
       });
 
       it('should return list of authorizations with unknown oauthMethod', async () => {
-        stubMethod($$.SANDBOX, AliasAccessor.prototype, 'getAll').returns([]);
         const auths = await AuthInfo.listAllAuthorizations((orgAuth) => orgAuth.error === 'FAIL!');
         expect(auths).to.deep.equal([
           {
             aliases: [],
             configs: [],
-            username: 'espresso@coffee.com',
-            orgId: '00DAuthInfoTest_orgId',
-            instanceUrl: 'https://mydevhub.localhost.internal.salesforce.com:6109',
+            username: testOrg.username,
+            orgId: testOrg.orgId,
+            instanceUrl: testOrg.instanceUrl,
             isExpired: 'unknown',
             accessToken: undefined,
             oauthMethod: 'unknown',
@@ -1849,15 +1574,15 @@ describe('AuthInfo', () => {
       });
 
       it('should return list of authorizations with unknown oauthMethod and alias', async () => {
-        stubMethod($$.SANDBOX, AliasAccessor.prototype, 'getAll').returns(['MyAlias']);
+        $$.stubAliases({ MyAlias: testOrg.username });
         const auths = await AuthInfo.listAllAuthorizations();
         expect(auths).to.deep.equal([
           {
             aliases: ['MyAlias'],
             configs: [],
-            username: 'espresso@coffee.com',
-            orgId: '00DAuthInfoTest_orgId',
-            instanceUrl: 'https://mydevhub.localhost.internal.salesforce.com:6109',
+            username: testOrg.username,
+            orgId: testOrg.orgId,
+            instanceUrl: testOrg.instanceUrl,
             isExpired: 'unknown',
             accessToken: undefined,
             oauthMethod: 'unknown',
@@ -1868,7 +1593,7 @@ describe('AuthInfo', () => {
     });
   });
 
-  describe('parseSfdxAuthUrl()', () => {
+  describe('parseSfdxAuthUrl', () => {
     it('should parse the correct url with no client secret', () => {
       const options = AuthInfo.parseSfdxAuthUrl(
         'force://PlatformCLI::5Aep861_OKMvio5gy8xCNsXxybPdupY9fVEZyeVOvb4kpOZx5Z1QLB7k7n5flEqEWKcwUQEX1I.O5DCFwjlYUB.@test.my.salesforce.com'
@@ -1921,23 +1646,25 @@ describe('AuthInfo', () => {
       }
     });
   });
-  describe('Handle User Get Errors', () => {
-    let authCodeConfig: any;
-    beforeEach(async () => {
+
+  describe('Handle User HTTP Get Errors', () => {
+    let authCodeConfig: { authCode: string; loginUrl: string };
+
+    beforeEach(() => {
       authCodeConfig = {
-        authCode: testMetadata.authCode,
-        loginUrl: testMetadata.loginUrl,
-      };
-      const authResponse = {
-        access_token: testMetadata.accessToken,
-        instance_url: testMetadata.instanceUrl,
-        id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
-        refresh_token: testMetadata.refreshToken,
+        authCode: testOrg.authcode,
+        loginUrl: testOrg.loginUrl,
       };
 
       // Stub the http requests (OAuth2.requestToken() and the request for the username)
-      _postParmsStub.returns(Promise.resolve(authResponse));
+      postParamsStub.resolves({
+        access_token: testOrg.accessToken,
+        instance_url: testOrg.instanceUrl,
+        id: '00DAuthInfoTest_orgId/005AuthInfoTest_userId',
+        refresh_token: testOrg.refreshToken,
+      });
     });
+
     it('user get returns 403 with body of json array', async () => {
       const responseBody = {
         statusCode: 403,
@@ -1948,7 +1675,7 @@ describe('AuthInfo', () => {
           },
         ]),
       };
-      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest').returns(Promise.resolve(responseBody));
+      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest').resolves(responseBody);
       try {
         await AuthInfo.create({ oauth2Options: authCodeConfig });
         assert(false, 'should throw');
@@ -1956,6 +1683,7 @@ describe('AuthInfo', () => {
         expect(err).to.have.property('message').to.include('The REST API is not enabled for this Organization');
       }
     });
+
     it('user get returns 403 with body of json map', async () => {
       const responseBody = {
         statusCode: 403,
@@ -1964,7 +1692,7 @@ describe('AuthInfo', () => {
           errorCode: 'RESTAPINOTENABLED',
         }),
       };
-      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest').returns(Promise.resolve(responseBody));
+      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest').resolves(responseBody);
       try {
         await AuthInfo.create({ oauth2Options: authCodeConfig });
         assert(false, 'should throw');
@@ -1972,12 +1700,13 @@ describe('AuthInfo', () => {
         expect(err).to.have.property('message').to.include('The REST API is not enabled for this Organization');
       }
     });
+
     it('user get returns 403 with string body', async () => {
       const responseBody = {
         statusCode: 403,
         body: 'The REST API is not enabled for this Organization',
       };
-      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest').returns(Promise.resolve(responseBody));
+      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest').resolves(responseBody);
       try {
         await AuthInfo.create({ oauth2Options: authCodeConfig });
         assert(false, 'should throw');
@@ -1985,11 +1714,10 @@ describe('AuthInfo', () => {
         expect(err).to.have.property('message').to.include('The REST API is not enabled for this Organization');
       }
     });
+
     it('user get returns server error with no body', async () => {
-      const responseBody = {
-        statusCode: 500,
-      };
-      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest').returns(Promise.resolve(responseBody));
+      const responseBody = { statusCode: 500 };
+      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest').resolves(responseBody);
       try {
         await AuthInfo.create({ oauth2Options: authCodeConfig });
         assert(false, 'should throw');
@@ -1997,12 +1725,13 @@ describe('AuthInfo', () => {
         expect(err).to.have.property('message').to.include('UNKNOWN');
       }
     });
+
     it('user get returns server error with html body', async () => {
       const responseBody = {
         statusCode: 500,
         body: '<html lang=""><body>Server error occurred, please contact Salesforce Support if the error persists</body></html>',
       };
-      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest').returns(Promise.resolve(responseBody));
+      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest').resolves(responseBody);
       try {
         await AuthInfo.create({ oauth2Options: authCodeConfig });
         assert(false, 'should throw');
@@ -2011,23 +1740,18 @@ describe('AuthInfo', () => {
       }
     });
   });
-});
-
-describe('align srcatch orgs with devhub', () => {
-  let adminTestData: MockTestOrgData;
-  let user1: MockTestOrgData;
-
-  beforeEach(async () => {
-    adminTestData = new MockTestOrgData();
-    user1 = new MockTestOrgData();
-  });
 
   describe('getDevHubAuthInfos', () => {
-    it('should not find a dev hub when no authInfos exist', async () => {
-      stubMethod($$.SANDBOX, AuthInfo, 'listAllAuthorizations').callsFake(async (): Promise<string[]> => {
-        return Promise.resolve([]);
-      });
+    let adminTestData: MockTestOrgData;
+    let user1: MockTestOrgData;
 
+    beforeEach(async () => {
+      adminTestData = new MockTestOrgData();
+      user1 = new MockTestOrgData();
+    });
+
+    it('should not find a dev hub when no authInfos exist', async () => {
+      stubMethod($$.SANDBOX, AuthInfo, 'listAllAuthorizations').resolves([]);
       const result = await AuthInfo.getDevHubAuthInfos();
       expect(result).to.have.lengthOf(0);
     });
@@ -2047,6 +1771,14 @@ describe('align srcatch orgs with devhub', () => {
   });
 
   describe('identifyPossibleScratchOrgs', () => {
+    let adminTestData: MockTestOrgData;
+    let user1: MockTestOrgData;
+
+    beforeEach(async () => {
+      adminTestData = new MockTestOrgData();
+      user1 = new MockTestOrgData();
+    });
+
     it('should not update org - no dev hubs', async () => {
       await $$.stubAuths(adminTestData, user1);
 
@@ -2123,5 +1855,92 @@ describe('align srcatch orgs with devhub', () => {
       expect(queryScratchOrgStub.callCount).to.be.equal(1);
       expect(authInfoSaveStub.callCount).to.be.equal(1);
     });
+  });
+
+  describe('determineIfDevHub', () => {
+    it('should return true if request succeeds', async () => {
+      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest').resolves({
+        statusCode: 200,
+        body: JSON.stringify([]),
+      });
+      const authInfo = await AuthInfo.create({ username: testOrg.username });
+      // @ts-expect-error because private method
+      expect(await authInfo.determineIfDevHub(testOrg.instanceUrl, testOrg.accessToken)).to.be.true;
+    });
+
+    it('should return false if request returns 400', async () => {
+      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest').resolves({
+        statusCode: 400,
+        body: JSON.stringify([]),
+      });
+      const authInfo = await AuthInfo.create({ username: testOrg.username });
+      // @ts-expect-error because private method
+      expect(await authInfo.determineIfDevHub(testOrg.instanceUrl, testOrg.accessToken)).to.be.false;
+    });
+
+    it('should return false if request fails', async () => {
+      stubMethod($$.SANDBOX, Transport.prototype, 'httpRequest').throws();
+      const authInfo = await AuthInfo.create({ username: testOrg.username });
+      // @ts-expect-error because private method
+      expect(await authInfo.determineIfDevHub(testOrg.instanceUrl, testOrg.accessToken)).to.be.false;
+    });
+  });
+
+  describe('loadDecryptedAuthFromConfig', () => {
+    const expectedErrorName = 'NamedOrgNotFoundError';
+    it('should throw error if no auth file is found', async () => {
+      try {
+        const authInfo = await AuthInfo.create({ username: testOrg.username });
+        // @ts-expect-error because private method
+        await authInfo.loadDecryptedAuthFromConfig('DOES_NOT_EXIST');
+        assert.fail(`should have thrown error with name: ${expectedErrorName}`);
+      } catch (e) {
+        expect(e).to.have.property('name', expectedErrorName);
+      }
+    });
+  });
+});
+
+describe('AuthInfo No fs mock', () => {
+  const $$ = testSetup();
+  const TEST_KEY = {
+    service: 'sfdx',
+    account: 'local',
+    key: '8e8fd1e6dc06a37bf420898dbc3ee35c',
+  };
+
+  beforeEach(() => {
+    // Testing crypto functionality, so restore global stubs.
+    $$.SANDBOXES.CRYPTO.restore();
+    $$.SANDBOXES.CONFIG.restore();
+    $$.SANDBOXES.ORGS.restore();
+
+    stubMethod($$.SANDBOX, Crypto.prototype, 'getKeyChain').callsFake(() =>
+      Promise.resolve({
+        setPassword: () => Promise.resolve(),
+        getPassword: (data: JsonMap, cb: (val1: AnyJson, key: string) => {}) => cb(null, TEST_KEY.key),
+      })
+    );
+  });
+
+  it('missing config', async () => {
+    const expectedErrorName = 'NamedOrgNotFoundError';
+    try {
+      await AuthInfo.create({ username: 'does_not_exist@gb.com' });
+      assert.fail(`should have thrown error with name: ${expectedErrorName}`);
+    } catch (e) {
+      expect(e).to.have.property('name', expectedErrorName);
+    }
+  });
+
+  it('invalid devhub username', async () => {
+    const expectedErrorName = 'NamedOrgNotFoundError';
+    try {
+      await AuthInfo.create({ username: 'does_not_exist@gb.com', isDevHub: true });
+      assert.fail(`should have thrown error with name: ${expectedErrorName}`);
+    } catch (e) {
+      expect(e).to.have.property('name', expectedErrorName);
+      expect(e).to.have.property('message', 'No authorization information found for does_not_exist@gb.com.');
+    }
   });
 });
